### 2020-10-30 6e805db
Update news after tag publication
