+++ .github/workflows/R-CMD-check.yaml
@@ -29,6 +29,7 @@ jobs:
 
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      R_CHECK_FORCE_SUGGESTS: false
       RSPM: ${{ matrix.config.rspm }}
 
     steps:
