### 2020-10-30 c576377
_R_CHECK_FORCE_SUGGESTS_ change in github workflow
