+++ .github/workflows/R-CMD-check.yaml
@@ -76,7 +76,7 @@ jobs:
           _R_CHECK_CRAN_INCOMING_REMOTE_: false
           _R_CHECK_FORCE_SUGGESTS_: false
           R_CHECK_DONTTEST_EXAMPLES: false
-        run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran"), error_on = "warning", check_dir = "check")
+        run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran", "--no-tests"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
 
       - name: Upload check results
