+++ .github/workflows/R-CMD-check.yaml
@@ -30,7 +30,7 @@ jobs:
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       _R_CHECK_FORCE_SUGGESTS_: false
-      R_CHECK_DONTTEST_EXAMPLES: false
+      _R_CHECK_DONTTEST_EXAMPLES: false
       RSPM: ${{ matrix.config.rspm }}
 
     steps:
@@ -75,7 +75,7 @@ jobs:
         env:
           _R_CHECK_CRAN_INCOMING_REMOTE_: false
           _R_CHECK_FORCE_SUGGESTS_: false
-          R_CHECK_DONTTEST_EXAMPLES: false
+          _R_CHECK_DONTTEST_EXAMPLES: false
         run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran", "--no-tests"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
 
