### 2020-11-02 2b1c2f0
Setting donttest examples as a global argument.
