+++ .github/workflows/R-CMD-check.yaml
@@ -30,8 +30,9 @@ jobs:
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       _R_CHECK_FORCE_SUGGESTS_: false
-      _R_CHECK_DONTTEST_EXAMPLES: false
       RSPM: ${{ matrix.config.rspm }}
+      global:
+        _R_CHECK_DONTTEST_EXAMPLES_: false
 
     steps:
       - uses: actions/checkout@v2
@@ -75,9 +76,10 @@ jobs:
         env:
           _R_CHECK_CRAN_INCOMING_REMOTE_: false
           _R_CHECK_FORCE_SUGGESTS_: false
-          _R_CHECK_DONTTEST_EXAMPLES: false
         run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran", "--no-tests"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
+        global:
+          _R_CHECK_DONTTEST_EXAMPLES_: false
 
       - name: Upload check results
         if: failure()
