### 2020-11-05 #407 @karthik
### Editor checks:

---

#### Editor comments

:wave: @hannahlowens. Here are some general checks from the goodpractice package. As you resolve these issues, you can run `goodpractice::gp()` yourself to make sure they are resolved. I'll start looking for reviewers now but if you have suggestions for people who can objectively review without a conflict of interest, I'm open to suggestions.

I also noticed there are no tests. Please add test coverage for all your functions. Let us know if you need help or pointers to get started.

```

It is good practice to

    R/BIENtableCleanup.R:19:NA
    R/BIENtableCleanup.R:20:NA
    R/BIENtableCleanup.R:21:NA
    R/BIENtableCleanup.R:22:NA
    R/BIENtableCleanup.R:23:NA
    ... and 833 more lines

  ✖ use '<-' for assignment instead of '='. '<-' is the
    standard, and R users and developers are used it and it is easier
    to read your code for them if you use '<-'.

    R/occCitation.R:89:11
    R/occCitation.R:90:13
    R/occCitation.R:91:11
    R/occCitation.R:92:15

    inst/extdata/sampleScript.R:79:1
    R/data.R:3:1
    R/data.R:4:1
    R/data.R:11:1
    R/data.R:13:1
    ... and 149 more lines

  ✖ omit trailing semicolons from code lines. They are not
    needed and most R coding standards forbid them

    inst/extdata/sampleScript.R:1:17
    inst/extdata/sampleScript.R:21:47
    inst/extdata/sampleScript.R:27:16
    inst/extdata/sampleScript.R:37:59
    inst/extdata/sampleScript.R:44:59
    ... and 17 more lines

  ✖ avoid calling setwd(), it changes the global environment.
    If you need it, consider using on.exit() to restore the working
    directory.

    R/getGBIFpoints.R:34:11
    R/getGBIFpoints.R:71:3
    R/getGBIFpoints.R:88:3
    R/occQuery.R:80:11
    R/occQuery.R:165:7
    ... and 1 more lines

    vapply() instead.

    R/occCite_plotting.R:28:20
    R/occCite_plotting.R:171:28

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...),
    1:NROW(...) and 1:NCOL(...) expressions. They are error prone and
    result 1:0 if the expression on the right hand side is zero. Use
    seq_len() or seq_along() instead.

    R/gbifRetriever.R:36:13
    R/occCitation.R:125:17
    R/occCitation.R:131:17
    R/occCitation.R:137:17
    R/occCitation.R:144:17
    ... and 9 more lines

  ✖ not import packages as a whole, as this can cause name
    clashes between the imported packages. Instead, import only the
    specific functions you need.
  ✖ fix this R CMD check ERROR: Package suggested but not
    available: 'RefManageR' The suggested packages are required for a
    complete check. Checking can be attempted without them by setting
    the environment variable _R_CHECK_FORCE_SUGGESTS_ to a false value.
    See section 'The DESCRIPTION file' in the 'Writing R Extensions'
    manual.
```
---

Reviewers:
Due date:
