### 2020-11-06 #407 @karthik
Hello Hannah. By tests, I mean explicit tests for each functions in the tests folder. Examples are great and can be turned into tests.  What you have are more along the lines of informal tests. For an introduction to writing tests, this chapter from the R packages book https://r-pkgs.org/tests.html might be useful.
