### 2020-11-09 1c517fa
Wrote BIENtableCleanup out of the package. It was reduntant and made error checking annoying. Also started writing formal tests.
