+++ man/taxonRectification.Rd
@@ -25,6 +25,6 @@ warnings if the name is invalid.
 }
 \examples{
 #Inputting a taxonomic name and specifying what taxonomic sources you want to search
-taxonRectification(taxName = "Buteo buteo hartedi", datasources = 'NCBI');
+taxonRectification(taxName = "Buteo buteo hartedi", datasources = 'NCBI')
 
 }
