+++ NAMESPACE
@@ -2,7 +2,6 @@
 
 S3method(print,occCiteCitation)
 S3method(summary,occCiteData)
-export(BIENtableCleanup)
 export(GBIFLoginManager)
 export(getBIENpoints)
 export(getGBIFpoints)
