### 2020-11-09 4e821e0
Move R cmd check to top of readme.
