+++ readme.md
@@ -1,3 +1,9 @@
+<!-- badges: start -->
+
+[![R build
+status](https://github.com/hannahlowens/occCite/workflows/R-CMD-check/badge.svg)](https://github.com/hannahlowens/occCite/actions)
+<!-- badges: end -->
+
 Introduction
 ============
 
@@ -140,7 +146,7 @@ summary(mySimpleOccCiteObject)
 ```
 
     ##  
-    ##  OccCite query occurred on: 30 October, 2020
+    ##  OccCite query occurred on: 09 November, 2020
     ##  
     ##  User query type: User-supplied list of taxa.
     ##  
@@ -309,7 +315,7 @@ summary(myOldOccCiteObject)
 ```
 
     ##  
-    ##  OccCite query occurred on: 30 October, 2020
+    ##  OccCite query occurred on: 09 November, 2020
     ##  
     ##  User query type: User-supplied list of taxa.
     ##  
@@ -407,7 +413,7 @@ summary(myPhyOccCiteObject)
 ```
 
     ##  
-    ##  OccCite query occurred on: 30 October, 2020
+    ##  OccCite query occurred on: 09 November, 2020
     ##  
     ##  User query type: User-supplied phylogeny.
     ##  
@@ -473,7 +479,7 @@ print(myPhyOccCitations)
     ## Bentley A (2020). KUBI Ichthyology Collection. Version 17.67. University of Kansas Biodiversity Institute. https://doi.org/10.15468/mgjasg. Accessed via GBIF on 2019-07-04.
     ## Bentley A (2020). KUBI Ichthyology Tissue Collection. Version 18.55. University of Kansas Biodiversity Institute. https://doi.org/10.15468/jmsnwg. Accessed via GBIF on 2019-07-04.
     ## Casassovici A, Brosens D (2020). Diveboard - Scuba diving citizen science observations. Version 54.30. Diveboard. https://doi.org/10.15468/tnjrgy. Accessed via GBIF on 2019-07-04.
-    ## Catania D, Fong J (2020). CAS Ichthyology (ICH). Version 150.246. California Academy of Sciences. https://doi.org/10.15468/efh2ib. Accessed via GBIF on 2019-07-04.
+    ## Catania D, Fong J (2020). CAS Ichthyology (ICH). Version 150.247. California Academy of Sciences. https://doi.org/10.15468/efh2ib. Accessed via GBIF on 2019-07-04.
     ## Cauquil P, Barde J (2011). observe_tuna_bycatch_ecoscope. IRD - Institute of Research for Development. https://doi.org/10.15468/23m361. Accessed via GBIF on 2019-07-04.
     ## Chamberlain, S., Barve, V., Mcglinn, D., Oldoni, D., Desmet, P., Geffert, L., Ram, K. (2020). rgbif: Interface to the Global Biodiversity Information Facility API. R package version 3.3.0. https://CRAN.R-project.org/package=rgbif.
     ## Chamberlain, S., Boettiger, C. (2017). R Python, and Ruby clients for GBIF species occurrence data. PeerJ PrePrints.
@@ -486,11 +492,9 @@ print(myPhyOccCitations)
     ## Frable B (2019). SIO Marine Vertebrate Collection. Version 1.7. Scripps Institution of Oceanography. https://doi.org/10.15468/ad1ovc. Accessed via GBIF on 2019-07-04.
     ## González Acosta A F, Comisión nacional para el conocimiento y uso de la biodiversidad C (2020). Ampliación de la base de datos de la ictiofauna insular del Golfo de California. Version 1.7. Comisión nacional para el conocimiento y uso de la biodiversidad. https://doi.org/10.15468/p5ovq7. Accessed via GBIF on 2019-07-04.
[…]
