### 2020-11-09 d6af367
test file for getBIENpoints
