### 2020-11-10 8b656a4
gbifRetriever tests written
