+++ R/GBIFLoginManager.R
@@ -64,7 +64,7 @@ GBIFLoginManager <- function(user = NULL, email = NULL, pwd = NULL) {
   return(loginInstance)
 }
 
-#Functions for checking for login information in system environment(adapted from \code{\link{occ_download}} in \code{\link{rgbif}})
+#Functions for checking for login information in system environment(adapted from occ_download in rgbif)
 check_user <- function(x) {
   z <- if (is.null(x)) Sys.getenv("GBIF_USER", "") else x
   if (z == "") getOption("gbif_user", stop("supply a username")) else z
