### 2020-11-11 cd3b879
wrote tests for GBIFtableCleanup
