+++ man/gbifRetriever.Rd
@@ -12,7 +12,7 @@ gbifRetriever(taxon = NULL)
 \value{
 A list of lists containing (1) a data frame of occurrence
 data; (2) GBIF search metadata for every GBIF download in the
-specified directory.
+specified directory. (3) the raw occurrence data from GBIF.
 }
 \description{
 Searches for the most recent instance of a data set
