### 2020-11-12 613c5e3
Wrote tests for occCitation function
