+++ tests/testthat/test-occCitation.R
@@ -39,3 +39,25 @@ test_that("outputs for GBIF results are as expected", {
   expect_true("Number of Occurrences" %in% names(testResults$occCitationResults[[1]]))
   expect_true(is.numeric(testResults$occCitationResults[[1]]$`Number of Occurrences`))
 })
+
+test_that("outputs for BIEN results are as expected", {
+  data("myOccCiteObject")
+  myOccCiteObject@occResults[[1]]$GBIF <- NULL
+  testResults <- occCitation(myOccCiteObject)
+
+  expect_true(class(testResults)=="occCiteCitation")
+
+  expect_true("occCitationResults" %in% names(testResults))
+  expect_equal(class(testResults$occCitationResults[[1]]), "data.frame")
+
+  expect_true("occSearch" %in% names(testResults$occCitationResults[[1]]))
+  expect_equal(class(testResults$occCitationResults[[1]]$occSearch), "character")
+  expect_true("Dataset Key" %in% names(testResults$occCitationResults[[1]]))
+  expect_equal(class(testResults$occCitationResults[[1]]$`Dataset Key`), "character")
+  expect_true("Citation" %in% names(testResults$occCitationResults[[1]]))
+  expect_equal(class(testResults$occCitationResults[[1]]$Citation), "character")
+  expect_true("Accession Date" %in% names(testResults$occCitationResults[[1]]))
+  expect_true(is.Date(as.Date(testResults$occCitationResults[[1]]$`Accession Date`)))
+  expect_true("Number of Occurrences" %in% names(testResults$occCitationResults[[1]]))
+  expect_true(is.numeric(testResults$occCitationResults[[1]]$`Number of Occurrences`))
+})
