### 2020-11-12 54de365
simplify tabulate.occResults (there's no need to ask for a species's name) and dependent fuctions that use it
