### 2020-11-12 bef0b2f
simplifying the names broke more than it fixed. I have reverted tabulate.occResults()
