### 2020-11-12 51947fb
Explicitly return map object
