### 2020-11-12 60da196
Wrote tests for mapp.occCite
