### 2020-11-12 00bd134
Tests for sumFig done.
