### 2020-11-20 5dd31c7
change "userSpecTaxonomicSources" slot in occCiteData to "userSpecTaxonomy"
