+++ R/occSum.R
@@ -29,10 +29,10 @@ summary.occCiteData <- function(object, ...) {
         sprintf("User query type: %s\n", x@userQueryType))
   }
 
-  if(!is.null(x@userSpecTaxonomicSources)){
+  if(!is.null(x@userSpecTaxonomy)){
     cat("\t\n",
         sprintf("Sources for taxonomic rectification: %s\n",
-                paste0(x@userSpecTaxonomicSources, collapse = ", ")), "\t\n")
+                paste0(x@userSpecTaxonomy, collapse = ", ")), "\t\n")
   }
 
   if(!is.null(x@cleanedTaxonomy)){
