+++ man/occCiteData-class.Rd
@@ -16,7 +16,7 @@ queries and data manipulation.
 user made their original taxonomic query based on a vector of taxon names
 or a phylogeny.}
 
-\item{\code{userSpecTaxonomicSources}}{A vector of type character that presents a
+\item{\code{userSpecTaxonomy}}{A vector of type character that presents a
 list of taxonomic sources for cleaning taxonomy of queries. This can be
 user-specified or default.}
 
