+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: 2.7.3
+pkgdown: 1.6.1
+pkgdown_sha: ~
+articles:
+  occCite_Vignette: occCite_Vignette.html
+last_built: 2020-11-20T14:44Z
+
