### 2020-11-23 74d81ea
New tests for getGBIFpoints and tabGBIF
