### 2020-11-23 73fb1c7
Getting rid of a couple of minor special-circumstance problems.
