### 2020-11-24 817c1a7
Cleaning up query date management and reporting.
