+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   occCite_Vignette: occCite_Vignette.html
-last_built: 2020-11-24T11:57Z
+last_built: 2020-11-24T12:57Z
 
