### 2020-11-24 c44bda1
Error checking the tests so they are robust to GBIF login failures.
