+++ tests/testthat/test-tabGBIF.R
@@ -4,8 +4,17 @@ library(occCite)
 
 test_that("data entered into tabGBIF is as expected", {
   skip_on_cran() # Requires GBIF login information
-  key <- rgbif::name_suggest(q="Protea cynaroides", rank='species')$data$key[1]
+  skip_if(nchar(Sys.getenv("GBIF_EMAIL")) < 1, "GBIF Login information not available")
+
+  test <- try(rgbif::occ_download(user=GBIFLogin@username,
+                                  email = GBIFLogin@email,
+                                  pwd = GBIFLogin@pwd,
+                                  rgbif::pred("catalogNumber", 217880)), silent = T)
+  skip_if(class(test) != 'occ_download', "GBIF login unsuccessful")
+
   GBIFLogin <- GBIFLoginManager()
+
+  key <- rgbif::name_suggest(q="Protea cynaroides", rank='species')$data$key[1]
   occD <- prevGBIFdownload(key, GBIFLogin)
   res <- rgbif::occ_download_get(key=occD, overwrite=TRUE,
                                  file.path(system.file('extdata/', package='occCite')))
@@ -15,8 +24,17 @@ test_that("data entered into tabGBIF is as expected", {
 
 test_that("verify occ_download_import results have expected columns", {
   skip_on_cran() # Requires GBIF login information
-  key <- rgbif::name_suggest(q="Protea cynaroides", rank='species')$data$key
+  skip_if(nchar(Sys.getenv("GBIF_EMAIL")) < 1, "GBIF Login information not available")
+
+  test <- try(rgbif::occ_download(user=GBIFLogin@username,
+                                  email = GBIFLogin@email,
+                                  pwd = GBIFLogin@pwd,
+                                  rgbif::pred("catalogNumber", 217880)), silent = T)
+  skip_if(class(test) != 'occ_download', "GBIF login unsuccessful")
+
   GBIFLogin <- GBIFLoginManager()
+
+  key <- rgbif::name_suggest(q="Protea cynaroides", rank='species')$data$key
   occD <- prevGBIFdownload(key, GBIFLogin)
   res <- rgbif::occ_download_get(key=occD, overwrite=TRUE,
                                  file.path(system.file('extdata/', package='occCite')))
@@ -35,8 +53,17 @@ test_that("verify occ_download_import results have expected columns", {
 
 test_that("tabGBIF results as expected", {
   skip_on_cran() # Requires GBIF login information
-  key <- rgbif::name_suggest(q="Protea cynaroides", rank='species')$data$key
+  skip_if(nchar(Sys.getenv("GBIF_EMAIL")) < 1, "GBIF Login information not available")
+
+  test <- try(rgbif::occ_download(user=GBIFLogin@username,
+                                  email = GBIFLogin@email,
+                                  pwd = GBIFLogin@pwd,
+                                  rgbif::pred("catalogNumber", 217880)), silent = T)
+  skip_if(class(test) != 'occ_download', "GBIF login unsuccessful")
+
   GBIFLogin <- GBIFLoginManager()
+
+  key <- rgbif::name_suggest(q="Protea cynaroides", rank='species')$data$key
   occD <- prevGBIFdownload(key, GBIFLogin)
   res <- rgbif::occ_download_get(key=occD, overwrite=TRUE,
                                  file.path(system.file('extdata/', package='occCite')))
