### 2020-11-24 72b8710
prevGBIFdownload tests
