### 2020-11-24 36463c7
renamed map.occCite to occCiteMap
