+++ README.Rmd
@@ -220,6 +220,6 @@ sumFig.occCite(myPhyOccCiteObject,
 occCite also contains tools to map search results. We can pass occCite objects to be mapped as an interactive html widget using `leaflet`. If there are multiple species in a single occCite object, occurrences will be color-coded by species. If you click on a single occurrence, you will be able to see the species, the coordinates, the date on which it was collected, the dataset to which it belongs, and the dataservice that provided the record.
 
 ```{r mapping search results, eval = F, echo = T}
-map.occCite(mySimpleOccCiteObject, cluster = T)
+occCiteMap(mySimpleOccCiteObject, cluster = T)
 ```
 
