### 2020-11-25 7a53e4b
sumFig() function is now a `plot` method for objects of class `occCiteData`. * `map.occCite()` has been renamed `occCiteMap()` to avoid confusion with existing method naming conventions.
