+++ man/occCiteMap.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/occCite_plotting.R
-\name{map.occCite}
-\alias{map.occCite}
+\name{occCiteMap}
+\alias{occCiteMap}
 \title{Generating a map of downloaded points}
 \usage{
-map.occCite(
+occCiteMap(
   occCiteData,
   species_map = "all",
   species_colors = NULL,
@@ -53,6 +53,6 @@ is `FALSE` and species_colors are not specified, random colors from the `RColorB
 }
 \examples{
 data(myOccCiteObject)
-map.occCite(myOccCiteObject, cluster = FALSE)
+occCiteMap(myOccCiteObject, cluster = FALSE)
 
 }
