### 2020-11-25 954aee6
More graceful dealing with na values in year.
