### 2020-11-25 8d87f38
error when sp name is NA is fixed.
