### 2020-11-25 #407 @hannahlowens
Hi @karthik et al.!

Tests are now implemented in the package.
