### 2020-11-25 87339c5
Updated for CRAN submission
