+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: occCite
 Type: Package
 Title: Querying and Managing Large Biodiversity Occurrence Datasets
-Version: 0.4.2
+Version: 0.4.5
 Authors@R: c(
   person(given = "Hannah L.", family = "Owens", ,"hannah.owens@gmail.com", role=c("aut","cre"),
     comment = c(ORCID = "0000-0003-0071-1745")),
@@ -56,6 +56,7 @@ Imports:
     tidyr,
     RPostgreSQL,
     DBI,
-    waffle
+    waffle,
+    testthat
 VignetteBuilder: knitr
 RoxygenNote: 7.1.1
