### 2020-11-25 711c48c
Trying to get automated checks functional.
