+++ .github/workflows/R-CMD-check.yaml
@@ -32,7 +32,7 @@ jobs:
       _R_CHECK_FORCE_SUGGESTS_: false
       RSPM: ${{ matrix.config.rspm }}
       global:
-        _R_CHECK_DONTTEST_EXAMPLES_: false
+        R_CHECK_DONTTEST_EXAMPLES: false
 
     steps:
       - uses: actions/checkout@v2
@@ -79,7 +79,7 @@ jobs:
         run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--as-cran", "--no-tests"), error_on = "warning", check_dir = "check")
         shell: Rscript {0}
         global:
-          _R_CHECK_DONTTEST_EXAMPLES_: false
+          R_CHECK_DONTTEST_EXAMPLES: false
 
       - name: Upload check results
         if: failure()
