### 2020-11-25 e1f3e27
Trying to fix a vignette build error in win-builder checks.
