### 2020-11-25 7feff99
Changed way of accessing package data folders...
