### 2020-11-25 902d048
Repair a no data error that occurs in the vignette and NOWHERE ELSE
