### 2020-11-27 7d386c8
Changed donttest to dontrun per CRAN recommendation (donttest examples were still being tested)
