+++ R/GBIFLoginManager.R
@@ -17,13 +17,13 @@
 #'
 #' @examples
 #' ## Inputting user particulars
-#'\donttest{
+#'\dontrun{
 #' myLogin <- GBIFLoginManager(user = "theWoman",
 #'                             email = "ireneAdler@@laScala.org",
 #'                             pwd = "sh3r")
 #'}
 #'
-#'\donttest{
+#'\dontrun{
 #' ## Can also be mined from your system environment
 #' myLogin <- GBIFLoginManager(user = NULL,
 #'                             email = NULL, pwd = NULL)
