+++ man/occQuery.Rd
@@ -55,7 +55,7 @@ Takes rectified list of specimens from
 \code{\link{rgbif}} with metadata.
 }
 \examples{
-\donttest{
+\dontrun{
 ##If you have already created a occCite object, and have not previously
 ##downloaded GBIF data.
 occQuery(x = myOccCiteObject,
