+++ man/getGBIFpoints.Rd
@@ -34,7 +34,7 @@ Downloads occurrence points and useful related information
 for processing within other occCite functions
 }
 \examples{
-\donttest{
+\dontrun{
 getGBIFpoints(taxon="Gadus morhua",
               GBIFLogin = myGBIFLogin,
               GBIFDownloadDirectory = NULL)
