+++ man/GBIFLoginManager.Rd
@@ -26,13 +26,13 @@ GBIF. You MUST ALREADY HAVE AN ACCOUNT at \href{http://gbif.org/}{GBIF}.
 }
 \examples{
 ## Inputting user particulars
-\donttest{
+\dontrun{
 myLogin <- GBIFLoginManager(user = "theWoman",
                             email = "ireneAdler@laScala.org",
                             pwd = "sh3r")
 }
 
-\donttest{
+\dontrun{
 ## Can also be mined from your system environment
 myLogin <- GBIFLoginManager(user = NULL,
                             email = NULL, pwd = NULL)
