### 2020-11-30 76d1460
First pass through using feedback from goodpractice R package
