+++ R/GBIFLoginManager.R
@@ -53,7 +53,11 @@ GBIFLoginManager <- function(user = NULL, email = NULL, pwd = NULL) {
   pwd <- check_pwd(pwd)
 
   #Test login
-  test <- try(rgbif::occ_download(user=user, email = email, pwd = pwd, rgbif::pred("catalogNumber", 217880)), silent = T)
+  test <- try(rgbif::occ_download(user=user,
+                                  email = email,
+                                  pwd = pwd,
+                                  rgbif::pred("catalogNumber", 217880)),
+              silent = T)
   if(class(test) != 'occ_download'){
     warning("GBIF user login data incorrect.\n")
     return(NULL)
