+++ R/data.R
@@ -1,7 +1,9 @@
 #' @title Results of an occCite search for *Protea cynaroides*
 #'
-#' @source Global Biodiversity Information Facility, GBIF (\url{https://www.gbif.org/}) and
-#' Botanical Information and Ecology Network, BIEN (\url{https://bien.nceas.ucsb.edu/bien/})
+#' @source Global Biodiversity Information Facility,
+#' GBIF (\url{https://www.gbif.org/}) and
+#' Botanical Information and Ecology Network,
+#' BIEN (\url{https://bien.nceas.ucsb.edu/bien/})
 #' data aggregators.
 #' @format An `occCiteData` object with the following slots:
 #' \describe{
@@ -10,8 +12,8 @@
 #'  \item{cleanedTaxonomy}{A data frame with results of taxonomic cleanup}
 #'  \item{occSources}{A vector of which databases were queried (i.e. GBIF and BIEN)}
 #'  \item{occCiteSearchDate}{When the search was made}
-#'  \item{occResults}{A list of length 1 named "Protea cynaroides". Contains a list of length
-#'  2 with results from each database, GBIF and BIEN}
+#'  \item{occResults}{A list of length 1 named "Protea cynaroides".
+#'  Contains a list of length 2 with results from each database, GBIF and BIEN}
 #' }
 #' @examples
 #' \donttest{
