### 2020-11-30 7b873f2
Batch two of goodpractice fixes.
