+++ R/data.R
@@ -10,7 +10,8 @@
 #'  \item{userQueryType}{What kind of query was made}
 #'  \item{userSpecTaxonomy}{A vector of taxonomic sources specified}
 #'  \item{cleanedTaxonomy}{A data frame with results of taxonomic cleanup}
-#'  \item{occSources}{A vector of which databases were queried (i.e. GBIF and BIEN)}
+#'  \item{occSources}{A vector of which databases were queried
+#'  (i.e. GBIF and BIEN)}
 #'  \item{occCiteSearchDate}{When the search was made}
 #'  \item{occResults}{A list of length 1 named "Protea cynaroides".
 #'  Contains a list of length 2 with results from each database, GBIF and BIEN}

[…]
