### 2020-11-30 34648d1
Roxygenizing documentation post line shortening.
