+++ man/occCiteCitation-class.Rd
@@ -12,9 +12,10 @@ queries.
 \section{Fields}{
 
 \describe{
-\item{\code{occCitationResults}}{The results of performing \code{\link{occCitation}} on a
-\code{\link{occCiteData}} object, stored as a named list, each of the items named
-after a searched taxon and containing a data frame with occurrence information.}
+\item{\code{occCitationResults}}{The results of performing
+\code{\link{occCitation}} on a \code{\link{occCiteData}} object,
+stored as a named list, each of the items named after a searched
+taxon and containing a data frame with occurrence information.}
 }}
 
 
