+++ readme.md
@@ -2,6 +2,10 @@
 
 [![R build
 status](https://github.com/hannahlowens/occCite/workflows/R-CMD-check/badge.svg)](https://github.com/hannahlowens/occCite/actions)
+[![cran
+version](https://www.r-pkg.org/badges/version/occCite)](https://cran.r-project.org/package=occCite)
+[![rstudio mirror
+downloads](https://cranlogs.r-pkg.org/badges/occCite)](https://github.com/r-hub/cranlogs.app)
 <!-- badges: end -->
 
 Introduction
@@ -87,6 +91,8 @@ mySimpleOccCiteObject <- occQuery(x = "Protea cynaroides",
                             checkPreviousGBIFDownload = T)
 ```
 
+    ## Error in is.nan(x): default method not implemented for type 'list'
+
 Here is what the GBIF results look like:
 
 ``` r
@@ -94,27 +100,7 @@ Here is what the GBIF results look like:
 head(mySimpleOccCiteObject@occResults$`Protea cynaroides`$GBIF$OccurrenceTable)
 ```
 
-    ##                name longitude  latitude day month year
-    ## 1 Protea cynaroides  18.40540 -33.95891   4     1 2015
-    ## 2 Protea cynaroides  18.42350 -33.96619  20     6 2019
-    ## 3 Protea cynaroides  22.99340 -34.05478  16     6 2019
-    ## 4 Protea cynaroides  18.40232 -34.08405   9     6 2019
-    ## 5 Protea cynaroides  19.44807 -34.52123  13     6 2019
-    ## 6 Protea cynaroides  18.39757 -34.07418   9     6 2019
-    ##                                   Dataset                           DatasetKey
-    ## 1 iNaturalist research-grade observations 50c9509d-22c7-4a22-a47d-8c48425ef4a7
-    ## 2 iNaturalist research-grade observations 50c9509d-22c7-4a22-a47d-8c48425ef4a7
-    ## 3 iNaturalist research-grade observations 50c9509d-22c7-4a22-a47d-8c48425ef4a7
-    ## 4 iNaturalist research-grade observations 50c9509d-22c7-4a22-a47d-8c48425ef4a7
-    ## 5 iNaturalist research-grade observations 50c9509d-22c7-4a22-a47d-8c48425ef4a7
-    ## 6 iNaturalist research-grade observations 50c9509d-22c7-4a22-a47d-8c48425ef4a7
-    ##   DataService
-    ## 1        GBIF
-    ## 2        GBIF
-    ## 3        GBIF
-    ## 4        GBIF
-    ## 5        GBIF
-    ## 6        GBIF
+    ## Error in head(mySimpleOccCiteObject@occResults$`Protea cynaroides`$GBIF$OccurrenceTable): object 'mySimpleOccCiteObject' not found
 
 And here are the BIEN results:
 
@@ -123,20 +109,7 @@ And here are the BIEN results:
 head(mySimpleOccCiteObject@occResults$`Protea cynaroides`$BIEN$OccurrenceTable)
 ```
 
-    ##                name longitude latitude day month year Dataset DatasetKey
-    ## 1 Protea cynaroides    22.875  -33.875  20     8 1973   SANBI       2249
-    ## 2 Protea cynaroides    25.125  -33.875   3     7 1934   SANBI       2249
-    ## 3 Protea cynaroides    20.375  -33.875  16     8 1952   SANBI       2249
-    ## 4 Protea cynaroides    21.375  -33.375  20     3 1947   SANBI       2249
-    ## 5 Protea cynaroides    20.875  -34.125  21     6 1987   SANBI       2249
[…]
