### 2021-01-08 958bef8
is.nan not behaving analgously to is.na or is.null
