### 2021-01-11 ad6ff6a
Addressing error when GBIF search returns no results that meet internal criteria for "legit" occurrence records.
