### 2021-01-11 4ee44b9
Making logic even more explicit.
