+++ R/taxonRectification.R
@@ -16,7 +16,8 @@
 #'
 #' @examples
 #' #Inputting taxonomic name and specifying what taxonomic sources to search
-#' taxonRectification(taxName = "Buteo buteo hartedi", datasources = 'NCBI')
+#' taxonRectification(taxName = "Buteo buteo hartedi",
+#'                    datasources = 'National Center for Biotechnology Information')
 #'
 #' @export
 #'
