### 2021-02-01 c116f63
Change double to single logic operators when appropriate--was throwing a warning in Windows.
