+++ cran-comments.md
@@ -11,26 +11,24 @@ This is a package update. In this version, I have:
 ## R CMD check results
 There were no WARNINGs. 
 
-There was 1 NOTE when using r-hub, win-builder devel, and win-builder release.
+There was 1 NOTE when using r-hub and local OS X.
 
-* checking CRAN incoming feasibility ... NOTE
-Maintainer: 'Hannah L. Owens <hannah.owens@gmail.com>'
-New submission
-
-Package was archived on CRAN
-
-CRAN repository db overrides:
-  X-CRAN-Comment: Archived on 2020-11-04 as check problems were not
-    corrected in time.
-    
-The check problems were corrected.
-
+* checking dependencies in R code ... NOTE
+  almost never needs to use ::: for its own objects:
+There are ::: calls to the package's namespace in its code. A package
+     almost never needs to use ::: for its own objects:
+     ‘GBIFtableCleanup’ ‘tabGBIF’
+     
+These internal functions are used by several package functions, and it was not feasible from an organizational perspective to put the helper functions in the same file as all the functions that use them.
+  
 There was 1 NOTE when using r-hub.
 
-* checking for future file timestamps ... NOTE
-unable to verify current time
-
-This appears to be a check function problem, not a problem with the package.
+* checking examples ... NOTE
+Examples with CPU (user + system) or elapsed time > 5s
+studyTaxonList 0.95   0.14    5.28
+               user system elapsed
+               
+This example uses an API. Usually it is faster than 5s.
 
 ## Downstream dependencies
 I also ran tools::check_packages_in_dir check on downstream dependencies of 
