### 2021-02-01 cc54549
finalized CRAN comments for new release.
