### 2021-02-02 8718858
Updated CRAN comments.
