+++ readme.md
@@ -8,11 +8,10 @@ version](https://www.r-pkg.org/badges/version/occCite)](https://cran.r-project.o
 downloads](https://cranlogs.r-pkg.org/badges/occCite)](https://github.com/r-hub/cranlogs.app)
 <!-- badges: end -->
 
-Introduction
-============
+# Introduction
 
-We have entered the age of data-intensive scientific discovery. As
-datasets increase in complexity and heterogeneity, we must preserve the
+We have entered the age of data-intensive scientific discovery. As data
+sets increase in complexity and heterogeneity, we must preserve the
 cycle of data citation from primary data sources to aggregating
 databases to research products and back to primary data sources. The
 citation cycle keeps science transparent, but it is also key to
@@ -40,7 +39,7 @@ for records. The results are appended to the `occCiteData` object, along
 with metadata on the search. Finally, the user can pass the
 `occCiteData` object to `occCitation`, which compiles citations for the
 primary providers, database aggregators, and `R` packages used to build
-the dataset.
+the data set.
 
 For an overview tutorial video of the package, see [our YouTube
 video](https://www.youtube.com/watch?v=7qSCULN_VjY&t=17s).
@@ -54,8 +53,7 @@ propose a new set of standards for citing primary biodiversity data in
 published research articles that provides due credit to contributors and
 allows them to track the use of their work. Keep checking back!
 
-Setup
-=====
+# Setup
 
 If you plan to query GBIF, you will need to provide them with your user
 login information. We have provided a dummy login below to show you the
@@ -72,11 +70,9 @@ GBIFLogin <- GBIFLoginManager(user = "occCiteTester",
                           pwd = "12345")
 ```
 
-Performing a simple search
-==========================
+# Performing a simple search
 
-The basics
-----------
+## The basics
 
 At its simplest, `occCite` allows you to search for occurrences for a
 single species. The taxonomy of the user-specified species will be
@@ -91,7 +87,13 @@ mySimpleOccCiteObject <- occQuery(x = "Protea cynaroides",
                             checkPreviousGBIFDownload = T)
 ```
 
-    ## Error in is.nan(x): default method not implemented for type 'list'
+    ## Warning in taxonRectification(taxName = targets[[count]], datasources =
+    ## datasources): Following sources not found in Global Names Index source list:
+    ## NCBI
+
+    ## Warning in taxonRectification(taxName = targets[[count]], datasources =
+    ## datasources): No valid taxonomic data sources supplied. Populating default list
+    ## from all available sources.
 
 Here is what the GBIF results look like:
 
@@ -100,7 +102,27 @@ Here is what the GBIF results look like:
 head(mySimpleOccCiteObject@occResults$`Protea cynaroides`$GBIF$OccurrenceTable)
 ```
 
-    ## Error in head(mySimpleOccCiteObject@occResults$`Protea cynaroides`$GBIF$OccurrenceTable): object 'mySimpleOccCiteObject' not found
+    ##                name longitude  latitude day month year
+    ## 1 Protea cynaroides  18.40540 -33.95891   4     1 2015
+    ## 2 Protea cynaroides  18.42350 -33.96619  20     6 2019
+    ## 3 Protea cynaroides  22.99340 -34.05478  16     6 2019
[…]
