+++ man/studyTaxonList.Rd
@@ -4,7 +4,10 @@
 \alias{studyTaxonList}
 \title{Study Taxon List}
 \usage{
-studyTaxonList(x = NULL, datasources = c("NCBI"))
+studyTaxonList(
+  x = NULL,
+  datasources = c("National Center for Biotechnology Information")
+)
 }
 \arguments{
 \item{x}{A phylogeny of class 'phylo' or a vector of class 'character'
