### 2021-03-11 bfd4a44
No longer assumes that names submitted as an occCite object relate back to GBIF keys.
