### 2021-03-11 ccafd5c
Further support for errors when a non-legit name is submitted in an occCite object.
