### 2021-03-24 2f3cf4d
Address error in summary() when only one source database is used.
