### 2021-03-24 6880197
Better handling of taxonomic mismatches. Search will no longer cancel when a mismatch is encountered.
