### 2021-04-15 8746535
Updated website, news, CRAN comments to prepare for new CRAN submission (and paper revision)
