### 2021-04-15 #407 @damianooldoni
I have never had working relationship with package authors.

- [X] **Examples** (that run successfully locally) for all exported functions

- Does not apply **Performance:** Any performance claims of the software been confirmed.

---

Well done! The package name fits the scope very well, it's catchy and easy to remember. Here below some comments to improve this package.

#### GitHub repo webpage

In https://github.com/hannahlowens/occCite I read in the Aout section:

> Querying database aggregators and citing primary sources of resulting occurrence points

Why do you speak about points? You mean occurrences in general which are not points, although I understand that we conceive them as points in the spacetime. Just replace _occurrence points_ with _occurrences_.

Some issues related to rOpenSci guidelines:

- README.md doesn't include a section about installing the package. Actually such section it's quite important as providing instructions about installation is one of the goals of any package's README. I suggest to add such info's in the existing _Setup_ section.
- rOpensci guidelines states that README should containg a _"brief demonstration usage"_. The README.md is way too long and duplicates information contained in the vignette. The README should point to other documentation sources via links to vignettes.
- Missing badges: test coverage, the badge for rOpenSci peer-review, a repostatus.org badge
- Missing citation information
- No code of conduct and contribution guidelines

Minor remarks:

- link [Global Names List](http://gni.globalnames.org/%7D) in README doesn't work.
- rgbif function `occ_search` has a hard limit of 100000 records, not 200000 as stated in README.md. At least this is what it's written in documentation (`?occ_search`) and how it works (rgbif version: 3.5.2):

```r
rgbif::occ_search(taxonKey = 1, limit = 200000)
Error: Max offset of 100001 exceeded: 99900 + 300
```

- Vignette is very long. I suggest to split it in two. The advanced features can be easily put in a second vignette called _Advanced_
- Vignette title is too long. Actual title: An overview of executing searches and generating citations using occCite. Suggested title: _Execute occurrence searches and generate citations_
- Code chunks in https://hannahlowens.github.io/occCite/articles/occCite_Vignette.html#loading-data-from-previous-gbif-searches return errors on website. To be fixed
- While running all chunks of Rmd vignette, I get 7 times the following warning:

```r
1: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
2: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
3: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
4: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
5: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
6: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
7: In utils::citation(pkg) :
  no date field in DESCRIPTION file of package ‘occCite’
```
I think you should get this fixed: occCite is about citation after all. To get an idea, if I try the same commando with `rgbif` package the following is returned:

```r
> utils::citation("rgbif")

Chamberlain S, Barve V, Mcglinn D, Oldoni D, Desmet P, Geffert L, Ram K (2021). _rgbif: Interface
to the Global Biodiversity Information Facility API_. R package version 3.5.2, <URL:
https://CRAN.R-project.org/package=rgbif>.

Chamberlain S, Boettiger C (2017). “R Python, and Ruby clients for GBIF species occurrence data.”
_PeerJ PrePrints_. <URL: https://doi.org/10.7287/peerj.preprints.3304v1>.

To see these entries in BibTeX format, use 'print(<citation>, bibtex=TRUE)', 'toBibtex(.)', or set
'options(citation.bibtex.max=999)'.
```

This remark is related to the missing citation information in README mentioned above.

### Code style

Personally I like tidyverse code style. For example, about comments style, I would add a space between the dash `#` and the comment as described in https://style.tidyverse.org/syntax.html#comments. And I would avoid to end up comments with a dot as I found in L23 of occCitation.R

To automatize the restyling `usethis::use_tidy_style()` can help a lot.

Please consider also to split strings longer than 80 characters. For example, L25 of `occCitation()`:

```r
warning("Input x is not of class 'occCiteData'. Input x must be result of a studyTaxonList() search.\n")
```

### Functions

#### GBIFLoginManager()

Basic but useful function! Nice. My only concern here is that this function silently(!) triggers a GBIF download of 109 occurrences every time it's called by a user, just to test if the credentials are correct. There are two problems doing this, both from user and GBIF side. As user I don't like to have my download page full of dummy downloads silently created every time I use this function. The triggering of downloads is also not mentioned in function documentation: I wonder what GBIF users will start to think when they notice such downloads: will they contact GBIF about it thinking about an error at GBIF side? :scream:
From GBIF side maybe it's not a big problem to have so many dummy downloads, but still, I would like to know whether you asked  GBIF about it. Maybe do they have an idea about using a dummy login for tests purposes?

In any case, I strongly suggest to remove the test. A strong reason to remove it is that the user still gets an error by `rgbif::occ_download()`  if credentials are not valid :+1

```r
> occ_download(user="user",
               email = "email",
               pwd = "pwd",
               rgbif::pred("catalogNumber", 217880))
Error: 401 - Unauthorized
```

I understand very well you want to apply _defensive programming_ in your package! And that's great. However, as `GBIFLoginManager()` is not adding much more than grouping user, email and pwd within an  instance of class `GBIFLogin`, I don't find adding another level of _defense_ really useful, espcially taking into account the drawbacks described above.

A typo: in function comment text at L66, replace `#Populating an instance of class occCiteData` with  `#Populating an instance of class GBIFLogin`

#### getGBIFpoints()

Same remark as for gitHub webpage above. I would replace `points` with `occs`, everywhere. This package is all about occurrences and contains the abbreviation _occ_ in the title after all! :smile

I would also mention GBIF in the @description. A suggestion:

```r
#' @description Downloads GBIF occurrences and useful related information for
#' processing within other occCite functions
```

I see that `getGBIFpoints()` calls rgbif function `occ_download()` harvesting only occurrences with coordinates and without any geospatial issue by using rgbif predicates (L64-65):

```r
rgbif::pred("hasCoordinate", TRUE)
rgbif::pred("hasGeospatialIssue", FALSE)
```

I suggest also to add a predicate to retrieve only presences. Not doing it can be very dangerous and a quite typical and naive error while processing GBIF data as the typical user is not concerned about occurrences related to absences.

A more general remark/question: what if users want to add other filters or modify yours? If I am interested only in data from 2018 from Andorra, why should I get data from all the world from 1000 up to now?

This is the classic dilemma while programming high level functions: the balance between being user friendly and flexibility.

Personally I would allow users to make their own query by means of ellipsis `...`. In this way your function is powerful but at the same time user friendly: GBIF/rgbif experts can specify their own predicates while other users just benefit of the basic functionality and so everybody is happy :smile:

In any case, even if you decide to not change your code, still you need to document your choice! The best way is to do it in Details section (@details) of the function documentation.

Another remark streaming from the previous one: I see you limit your research on species as you search among species names only. At L41:

```r
key <- rgbif::name_suggest(q=cleanTaxon, rank='species')$data$key[1]
```

I find this quite a limit: why don't you allow the user to specify the rank itself by adding an argument  `rank` to the function? In this way you make your function much more useful. Again, to solve the dilemma I cited above, you can add a default value (`"species"`) to this new argument. This is also important as it can be applied for a multi-species search! You can just pass a vector of species to argument `taxon` or simply the parent name, e.g. `Gadus` to get all occurrences of Gadus spp.

I would also suggest to add (at least) a `kingdom` argument to avoid [hemihomony](https://en.m.wikipedia.org/wiki/Homonym_(biology)#Hemihomonyms). An alternative would be to allow the user to select which taxon he/she means if GBIF returns multiple taxa instead of choosing the first key (see chunk code above). In case of

Based on this, I would also change description of argument `taxon` (`A single species`) to  `a valid taxon name` and some examples more:

```r
# taxon can be a species (default)
getGBIFpoints(taxon="Gadus morhua",
              GBIFLogin = myGBIFLogin,
              GBIFDownloadDirectory = NULL)
# taxon can be something else, e.g. a genus
getGBIFpoints(taxon="Gadus",
              rank = "genus"
              GBIFLogin = myGBIFLogin,
              GBIFDownloadDirectory = NULL)
```

Still speaking about documentation, I see you can pass a phylogeny object to `taxon`, which is great, but this should be documented in `@param taxon` field in Roxygen documentation as well.

Your package is actually quite taxonomic oriented: you MUST specify a taxon/phylogeny. But what if the user wants to download all data from a country or a polygon for example? This is something I do quite often. Again, take in mind the balance between flexibility and user friendiness.

A typo: comment text at L34: replace `#File hygene` with `#File hygiene`.

#### prevGBIFdownload()

This function is quite important to avoid downloading again and again same data:

```r
prevGBIFdownload(taxonKey = "No key", GBIFLogin)
```

Why do you have to specify a default value to `taxonKey`? I don't see any need in your code to do so. And in case you really want to specify a default value, I suggest to use `NULL` instead of a dummy value like `"No key "` and adapt the code in case needed. This is a kind of convention accepted by the R community.

Let's say I had to trigger a download via rgbif instead of using your getGBIFpoints() function as I need to download all data of all species in a region instead of specifying a taxon/phylogeny. Well, it would be nice to allow me to still use the rest of the package functionalities, isn't? This can be another way to solve the _dilemma_ between user friendly code and flexibility. How to do it? you can allow me to get my GBIF download by passing a GBIF download key or the GBIF DOI to `getGBIFpoints()` and so to `prevGBIFdownload()`!

```r
`prevGBIFdownload(taxonKey, GBIFLogin, GBIFDownloadKey)`
```

Not only, as the downloads are public, the use of GBIFLogin is not really needed if the GBIF download is provided! For example, you can download [this download](https://www.gbif.org/occurrence/download/0252206-200613084148143), with key 0252206-200613084148143 and DOI https://doi.org/10.15468/dl.bptja4 isn't?

Again, this improves the function: up to now the user is limited to the most recent download of the same taxon/philogeny. But waht if the last download has been deleted or it is corrupted or is invalid for any other reason?

#### getBIENpoints()

Quite the same remarks as for getGBIFpoints()

#### occCitation()

I think you forgot to specify which function(s) you need to import from RPostgreSQL package. At L13 I see:

```r
#' @import RPostgreSQL
```

#### tabulate.occResults()

You wrote `@keywords Internal` instead of `@keywords internal`: so this function is in the documentation index. It's also good practice in rOpensci to add `#' @noRd` for internal functions. Same holds true for tidyverse style guide: see https://style.tidyverse.org/documentation.html#internal-functions. Now I can see the [documentation of this function](https://hannahlowens.github.io/occCite/reference/tabulate.occResults.html) online, which is strange for internal functions.

Why do you use `dplyr::mutate_if`  `dplyr::mutate` and `dplyr::bind_rows` instead of adding them to the @importFrom dplyr as you do for  `%>%` and `filter`? Not only, `filter` is imported but never used in  `tabulate.occResults()`

#### occCiteMap

Again, is there any conflict with other packages for using  `dplyr::filter` instead of  `filter` as you import it?

All  `leaflet` functions are not in @importFrom

#### All other functions

I think a in-depth documentation check should be done throughout all other functions.

The idea of unit-tests is to tests as much _units_ of code as possible, and doing it fast. But tests should not rely or at least as less as possible on other packages or databases.

Is there a reason to use `catalogNumber` 217880 as predicate for triggering a GBIF download?

I would just pass an existent GBIF download key where it's possible so that your code is more independent of other packages.

Running `devtools::test()` returns 10 skipped tests: `Reason: GBIF login unsuccessful`. This should not happen on my laptop as I have the environmental variables all set up! The problem is that sometimes you use `GBIFLogin` object without first creating it. Example: L20 of `test-GBIFTableCleanup.R`.

So, be careful while using `try()` as you don't know if the code stops due to a bug or due to the reason you have in mind. And in general, IMHO I wouldn't use `try()` in tests: tests are not there to be tried, but to succeed or to fail.
