### 2021-04-16 baded28
removed .Rd files for internal functions per ROpenSci reviewer recommendation
