+++ man/tabGBIF.Rd
@@ -1,25 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tabGBIF.R
-\name{tabGBIF}
-\alias{tabGBIF}
-\title{GBIF Table}
-\usage{
-tabGBIF(GBIFresults, taxon)
-}
-\arguments{
-\item{GBIFresults}{The results of a GBIF search that will be tabulated
-into a common format for an occCite object.}
-
-\item{taxon}{A single species name, for tracing/error checking purposes
-only.}
-}
-\value{
-A list of lists containing \enumerate{ \item a data frame of
-occurrence data  \item GBIF search metadata for every GBIF download in
-the specified directory.}
-}
-\description{
-Internal function--imports results from `occ_download_get()`
-and processes them into a table for an `occCiteData` object.
-}
-\keyword{internal}
