+++ man/GBIFtableCleanup.Rd
@@ -1,21 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/GBIFtableCleanup.R
-\name{GBIFtableCleanup}
-\alias{GBIFtableCleanup}
-\title{GBIF Table Cleanup}
-\usage{
-GBIFtableCleanup(GBIFtable)
-}
-\arguments{
-\item{GBIFtable}{A data.frame resulting from \code{\link{getGBIFpoints}} OR
-\code{\link{gbifRetriever}}.}
-}
-\value{
-A data.frame tidied for compatibility with
-\code{\link{getBIENpoints}} results.
-}
-\description{
-Forces occurrence table columns to a uniform
-format. This is an internal function for use by \code{\link{occQuery}}.
-}
-\keyword{internal}
