+++ DESCRIPTION
@@ -22,8 +22,8 @@ Description: Facilitates the gathering of biodiversity occurrence data
   from disparate sources. Metadata is managed throughout the process to facilitate 
   reporting and enhanced ability to repeat analyses.
 License: GPL-3
-URL: https://github.com/hannahlowens/occCite
-BugReports: https://github.com/hannahlowens/occCite/issues
+URL: hannahlowens.github.io/occCite/
+BugReports: github.com/hannahlowens/occCite/issues
 Encoding: UTF-8
 LazyData: true
 Language: en-US
