### 2021-04-16 b44b176
Cleaned up code style using usethis::use_tidy_style()
