+++ man/prevGBIFdownload.Rd
@@ -22,12 +22,16 @@ prepared download.
 }
 \examples{
 \dontrun{
-GBIFLogin <- GBIFLoginManager(user = "theWoman",
-               email = "ireneAdler@laScala.org",
-               pwd = "sh3r")
-taxKey <- rgbif::name_suggest(q= "Protea cynaroides", rank='species')$key[1]
-prevGBIFdownload(taxonKey = taxKey,
-              GBIFLogin = myGBIFLogin)
+GBIFLogin <- GBIFLoginManager(
+  user = "theWoman",
+  email = "ireneAdler@laScala.org",
+  pwd = "sh3r"
+)
+taxKey <- rgbif::name_suggest(q = "Protea cynaroides", rank = "species")$key[1]
+prevGBIFdownload(
+  taxonKey = taxKey,
+  GBIFLogin = myGBIFLogin
+)
 }
 
 }
