+++ man/getBIENpoints.Rd
@@ -21,7 +21,7 @@ information for processing within other occCite functions
 }
 \examples{
 \dontrun{
-getBIENpoints(taxon="Protea cynaroides")
+getBIENpoints(taxon = "Protea cynaroides")
 }
 
 }
