+++ man/studyTaxonList.Rd
@@ -32,16 +32,21 @@ well as warnings if there are invalid names.
 }
 \examples{
 ## Inputting a vector of taxon names
-studyTaxonList(x = c("Buteo buteo",
-                     "Buteo buteo hartedi",
-                     "Buteo japonicus"),
-                     datasources = c('National Center for Biotechnology Information'))
+studyTaxonList(
+  x = c(
+    "Buteo buteo",
+    "Buteo buteo hartedi",
+    "Buteo japonicus"
+  ),
+  datasources = c("National Center for Biotechnology Information")
+)
 
 ## Inputting a phylogeny
 phylogeny <- ape::read.nexus(
-     system.file("extdata/Fish_12Tax_time_calibrated.tre",
-     package = "occCite"))
+  system.file("extdata/Fish_12Tax_time_calibrated.tre",
+    package = "occCite"
+  )
+)
 phylogeny <- ape::extract.clade(phylogeny, 18)
-studyTaxonList(x = phylogeny, datasources = c('National Center for Biotechnology Information'))
-
+studyTaxonList(x = phylogeny, datasources = c("National Center for Biotechnology Information"))
 }
