### 2021-04-16 6f84396
Worked on making lines shorter.
