+++ R/GBIFLoginManager.R
@@ -37,17 +37,20 @@
 GBIFLoginManager <- function(user = NULL, email = NULL, pwd = NULL) {
   # Error checking inputs
   if (!is.null(user) & class(user) != "character") {
-    warning("Input user name is invalid; it must be a vector of class 'character'.\n")
+    warning(paste0("Input user name is invalid;\n",
+                   "it must be a vector of class 'character'.\n"))
     return(NULL)
   }
 
   if (!is.null(email) & class(email) != "character") {
-    warning("Input email is invalid; it must be a vector of class 'character'.\n")
+    warning(paste0("Input email is invalid;\n",
+                   "it must be a vector of class 'character'.\n"))
     return(NULL)
   }
 
   if (!is.null(pwd) & class(pwd) != "character") {
-    warning("Input password is invalid; it must be a vector of class 'character'.\n")
+    warning(paste0("Input password is invalid;\n",
+            "it must be a vector of class 'character'.\n"))
     return(NULL)
   }
 
