+++ R/gbifRetriever.R
@@ -33,7 +33,9 @@ gbifRetriever <- function(taxon = NULL) {
     return(NULL)
   }
 
-  paths <- list.files(getwd(), pattern = "\\d{7}-\\d{15}.zip", recursive = T, full.names = T)
+  paths <- list.files(getwd(),
+                      pattern = "\\d{7}-\\d{15}.zip",
+                      recursive = T, full.names = T)
   keys <- as.vector(stringr::str_match(paths, pattern = "\\d{7}-\\d{15}"))
   paths <- stringr::str_remove(paths, pattern = "\\d{7}-\\d{15}.zip")
 
