+++ NAMESPACE
@@ -28,10 +28,9 @@ importFrom(ggplot2,ggplot)
 importFrom(ggplot2,ggplot_build)
 importFrom(ggplot2,ggtitle)
 importFrom(ggplot2,scale_y_continuous)
-importFrom(ggplot2,theme)
-importFrom(ggplot2,theme_classic)
-importFrom(ggplot2,xlab)
-importFrom(ggplot2,ylab)
+importFrom(ggplot2theme,theme_classic)
+importFrom(ggplot2theme,xlab)
+importFrom(ggplot2theme,ylab)
 importFrom(methods,is)
 importFrom(methods,new)
 importFrom(rlang,.data)

[…]
