+++ README.Rmd
@@ -24,6 +24,8 @@ knitr::opts_knit$set(root.dir = paste0(path.package("occCite"), "/extdata/"))
 [![R build status](https://github.com/hannahlowens/occCite/workflows/R-CMD-check/badge.svg)](https://github.com/hannahlowens/occCite/actions)
 [![cran version](https://www.r-pkg.org/badges/version/occCite)](https://cran.r-project.org/package=occCite)
 [![rstudio mirror downloads](https://cranlogs.r-pkg.org/badges/occCite)](https://github.com/r-hub/cranlogs.app)
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Codecov test coverage](https://codecov.io/gh/hannahlowens/occCite/branch/main/graph/badge.svg)](https://codecov.io/gh/hannahlowens/occCite?branch=main)
 <!-- badges: end -->
 
 # Introduction 
