+++ man/studyTaxonList.Rd
@@ -48,5 +48,6 @@ phylogeny <- ape::read.nexus(
   )
 )
 phylogeny <- ape::extract.clade(phylogeny, 18)
-studyTaxonList(x = phylogeny, datasources = c("National Center for Biotechnology Information"))
+studyTaxonList(x = phylogeny,
+               datasources = c("National Center for Biotechnology Information"))
 }
