### 2021-04-16 d3ea747
Some slightly more sophisticated roxygen2 work.
