+++ man/getBIENpoints.Rd
@@ -13,7 +13,7 @@ getBIENpoints(taxon)
 A list containing \enumerate{ \item a data frame of occurrence data;
 \item a list containing: i notes on usage, ii bibtex citations,
 and iii acknowledgment information; \item a data frame containing
-the raw results of a query to `BIEN::BIEN_occurrence_species()`.}
+the raw results of a query to \code{BIEN::BIEN_occurrence_species()}.}
 }
 \description{
 Downloads occurrence points and useful related
