+++ readme.md
@@ -6,6 +6,11 @@ status](https://github.com/hannahlowens/occCite/workflows/R-CMD-check/badge.svg)
 version](https://www.r-pkg.org/badges/version/occCite)](https://cran.r-project.org/package=occCite)
 [![rstudio mirror
 downloads](https://cranlogs.r-pkg.org/badges/occCite)](https://github.com/r-hub/cranlogs.app)
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Codecov test
+coverage](https://codecov.io/gh/hannahlowens/occCite/branch/main/graph/badge.svg)](https://codecov.io/gh/hannahlowens/occCite?branch=main)
 <!-- badges: end -->
 
 # Introduction
@@ -83,17 +88,9 @@ verified using EOL and NCBI taxonomies by default.
 mySimpleOccCiteObject <- occQuery(x = "Protea cynaroides",
                             datasources = c("gbif", "bien"),
                             GBIFLogin = GBIFLogin, 
-                            GBIFDownloadDirectory = paste0(path.package("occCite"), "/extdata/"),
+                            GBIFDownloadDirectory = system.file('extdata/', package='occCite'),
                             checkPreviousGBIFDownload = T)
 ```
-
-    ## Warning in taxonRectification(taxName = targets[[count]], datasources =
-    ## datasources): Following sources not found in Global Names Index source list:
-    ## NCBI
-
-    ## Warning in taxonRectification(taxName = targets[[count]], datasources =
-    ## datasources): No valid taxonomic data sources supplied. Populating default list
-    ## from all available sources.
 
 Here is what the GBIF results look like:
 
@@ -154,29 +151,29 @@ summary(mySimpleOccCiteObject)
 ```
 
     ##  
-    ##  OccCite query occurred on: 05 March, 2021
+    ##  OccCite query occurred on: 16 April, 2021
     ##  
     ##  User query type: User-supplied list of taxa.
     ##  
-    ##  Sources for taxonomic rectification: NCBI
+    ##  Sources for taxonomic rectification: National Center for Biotechnology Information
     ##      
     ##  
     ##  Taxonomic cleaning results:     
     ## 
     ##          Input Name        Best Match
     ## 1 Protea cynaroides Protea cynaroides
-    ##                                                                                                                                     Taxonomic Databases w/ Matches
-    ## 1 National Center for Biotechnology Information; Freebase; Encyclopedia of Life; CU*STAR; uBio NameBank; Arctos; Open Tree of Life Reference Taxonomy; iNaturalist
+    ##                  Taxonomic Databases w/ Matches
+    ## 1 National Center for Biotechnology Information
     ##  
     ##  Sources for occurrence data: gbif, bien
     ##      
     ##             Species Occurrences Sources
-    ## 1 Protea cynaroides         828      15
+    ## 1 Protea cynaroides           0       0
     ##  
     ##  GBIF dataset DOIs:  
     ## 
-    ##             Species GBIF Access Date           GBIF DOI
-    ## 1 Protea cynaroides       2019-07-15 10.15468/dl.iqnra2
+    ##             Species GBIF Access Date GBIF DOI
+    ## 1 Protea cynaroides             <NA>     <NA>
 
 ## Simple citations
 
[…]
