+++ README.Rmd
@@ -26,6 +26,7 @@ knitr::opts_knit$set(root.dir = paste0(path.package("occCite"), "/extdata/"))
 [![rstudio mirror downloads](https://cranlogs.r-pkg.org/badges/occCite)](https://github.com/r-hub/cranlogs.app)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 [![Codecov test coverage](https://codecov.io/gh/hannahlowens/occCite/branch/main/graph/badge.svg)](https://codecov.io/gh/hannahlowens/occCite?branch=main)
+[![](https://badges.ropensci.org/407_status.svg)](https://github.com/ropensci/software-review/issues/407)
 <!-- badges: end -->
 
 # Introduction 
