+++ README.Rmd
@@ -29,204 +29,48 @@ knitr::opts_knit$set(root.dir = paste0(path.package("occCite"), "/extdata/"))
 [![](https://badges.ropensci.org/407_status.svg)](https://github.com/ropensci/software-review/issues/407)
 <!-- badges: end -->
 
-# Introduction 
+Please cite rgbif. Run the following to get the appropriate citation for the version you're using:
 
-We have entered the age of data-intensive scientific discovery. As data sets increase in complexity and heterogeneity, we must preserve the cycle of data citation from primary data sources to aggregating databases to research products and back to primary data sources. The citation cycle keeps science transparent, but it is also key to supporting primary providers by documenting the use of their data. The Global Biodiversity Information Facility and other data aggregators have made great strides in harvesting citation data from research products and linking them back to primary data providers. However, this only works if those that publish research products cite primary data sources in the first place. We developed `occCite`, a set of `R`-based tools for downloading, managing, and citing biodiversity data, to advance toward the goal of closing the data provenance cycle. These tools preserve links between occurrence data and primary providers once researchers download aggregated data, and facilitate the citation of primary data providers in research papers.
+```r
+citation(package = "rgbif")
+```
 
-The `occCite` work flow follows a three-step process. First, the user inputs one or more taxonomic names (or a phylogeny). `occCite` then rectifies those names by checking them against one or more taxonomic databases, which can be specified by the user (see the [Global Names List](http://gni.globalnames.org/})). The results of the taxonomic rectification are then kept in an `occCiteData` object in local memory. Next, `occCite` takes the `occCiteData` object and user-defined search parameters to query BIEN (through `rbien`) and/or GBIF(through `rGBIF`) for records. The results are appended to the `occCiteData` object, along with metadata on the search. Finally, the user can pass the `occCiteData` object to `occCitation`, which compiles citations for the primary providers, database aggregators, and `R` packages used to build the data set.
+To get started, see:
+
+* occCite introduction vignette (https://docs.ropensci.org/rgbif/articles/rgbif.html)
+* occCite advanced feature vignette 
+* Function reference (https://docs.ropensci.org/rgbif/reference/index.html): an overview of all `rgbif` functions.
+* Articles (https://docs.ropensci.org/rgbif/articles/index.html): vignettes/tutorials on how to download data, clean data, and work with taxonomic names.
+
+Check out the `rgbif` [paper][] for more information on this package and the sister [Python][pygbif], [Ruby][gbifrb], and [PHP][phpgbif] clients.
+
+Note: Maximum number of records you can get with `occ_search()` and `occ_data()` is 100,000. See https://www.gbif.org/developer/occurrence
+
+## Installation
+
+```{r eval=FALSE}
+install.packages("rgbif")
+```
+
+Or, install development version
+
+```{r eval=FALSE}
+pak::pkg_install("ropensci/rgbif")
+# OR
+install.packages("rgbif", repos="https://dev.ropensci.org")
+```
+
+```{r}
+library("rgbif")
+```
+
+
+## Meta
+
+* Please [report any issues or bugs](https://github.com/hannahlowens/occCite/issues).
+* License: GPL-3
+* Get citation information for `occCite` in R using `citation(package = 'occCite')`
+
+- - -
 
 For an overview tutorial video of the package, see [our YouTube video](https://www.youtube.com/watch?v=7qSCULN_VjY&t=17s).
-
[…]
