+++ pkgdown.yml
@@ -2,6 +2,7 @@ pandoc: 2.7.3
 pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
-  occCite_Vignette: occCite_Vignette.html
-last_built: 2021-04-16T12:19Z
+  Advanced: Advanced.html
+  Simple: Simple.html
+last_built: 2021-04-16T15:15Z
 
