+++ readme.md
@@ -14,810 +14,65 @@ coverage](https://codecov.io/gh/hannahlowens/occCite/branch/main/graph/badge.svg
 [![](https://badges.ropensci.org/407_status.svg)](https://github.com/ropensci/software-review/issues/407)
 <!-- badges: end -->
 
-# Introduction
-
-We have entered the age of data-intensive scientific discovery. As data
-sets increase in complexity and heterogeneity, we must preserve the
-cycle of data citation from primary data sources to aggregating
-databases to research products and back to primary data sources. The
-citation cycle keeps science transparent, but it is also key to
-supporting primary providers by documenting the use of their data. The
-Global Biodiversity Information Facility and other data aggregators have
-made great strides in harvesting citation data from research products
-and linking them back to primary data providers. However, this only
-works if those that publish research products cite primary data sources
-in the first place. We developed `occCite`, a set of `R`-based tools for
-downloading, managing, and citing biodiversity data, to advance toward
-the goal of closing the data provenance cycle. These tools preserve
-links between occurrence data and primary providers once researchers
-download aggregated data, and facilitate the citation of primary data
-providers in research papers.
-
-The `occCite` work flow follows a three-step process. First, the user
-inputs one or more taxonomic names (or a phylogeny). `occCite` then
-rectifies those names by checking them against one or more taxonomic
-databases, which can be specified by the user (see the [Global Names
-List](http://gni.globalnames.org/%7D)). The results of the taxonomic
-rectification are then kept in an `occCiteData` object in local memory.
-Next, `occCite` takes the `occCiteData` object and user-defined search
-parameters to query BIEN (through `rbien`) and/or GBIF(through `rGBIF`)
-for records. The results are appended to the `occCiteData` object, along
-with metadata on the search. Finally, the user can pass the
-`occCiteData` object to `occCitation`, which compiles citations for the
-primary providers, database aggregators, and `R` packages used to build
-the data set.
-
-For an overview tutorial video of the package, see [our YouTube
-video](https://www.youtube.com/watch?v=7qSCULN_VjY&t=17s).
-
-Future iterations of `occCite` will track citation data through the data
-cleaning process and provide a series of visualizations on raw query
-results and final datasets. It will also provide data citations in a
-format congruent with best-practice recommendations for large
-biodiversity datasets. Based on these data citation tools, we will also
-propose a new set of standards for citing primary biodiversity data in
-published research articles that provides due credit to contributors and
-allows them to track the use of their work. Keep checking back!
-
-# Setup
-
-If you plan to query GBIF, you will need to provide them with your user
-login information. We have provided a dummy login below to show you the
-format. *You will need to provide actual account information.* This is
-because you will actually be downloading *all* of the records available
-for the species using `occ_download()`, instead of getting results from
-`occ_search()`, which has a hard limit of 200,000 occurrences.
+Please cite occCite. Run the following to get the appropriate citation
[…]
