+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   Advanced: Advanced.html
   Simple: Simple.html
-last_built: 2021-04-16T15:26Z
+last_built: 2021-04-16T19:56Z
 
