+++ README.Rmd
@@ -29,21 +29,15 @@ knitr::opts_knit$set(root.dir = paste0(path.package("occCite"), "/extdata/"))
 [![](https://badges.ropensci.org/407_status.svg)](https://github.com/ropensci/software-review/issues/407)
 <!-- badges: end -->
 
+## Summary
+
+The `occCite` workflow follows a three-step process. First, the user inputs one or more taxonomic names (or a phylogeny). `occCite` then rectifies these names by checking them against one or more taxonomic databases, which can be specified by the user (see the [Global Names List](http://gni.globalnames.org/data_sources)). The results of the taxonomic rectification are then kept in an `occCiteData` object in local memory. Next, `occCite` takes the `occCiteData` object and user-defined search parameters to query BIEN (through `rbien`) and/or GBIF(through `rGBIF`) for records. The results are appended to the `occCiteData` object, along with metadata on the search. Finally, the user can pass the `occCiteData` object to `occCitation`, which compiles citations for the primary providers, database aggregators, and `R` packages used to build the dataset.
+
 Please cite occCite. Run the following to get the appropriate citation for the version you're using:
 
 ```{r}
 citation(package = "occCite")
 ```
-
-To get started, see:
-
-* occCite introduction vignette (https://hannahlowens.github.io/occCite/articles/Advanced.html)
-* occCite advanced feature vignette (https://docs.ropensci.org/rgbif/articles/rgbif.html)
-* Function reference https://hannahlowens.github.io/occCite/reference/index.html
-* YouTube tutorial (https://www.youtube.com/watch?v=7qSCULN_VjY&t=17s)
-
-The `occCite` manuscript is currently in review. Check back soon for more information on that!
-
 
 ## Installation
 
@@ -60,6 +54,15 @@ devtools::install_github("DeveloperName/PackageName")("hannahlowens/occCite")
 ```{r}
 library("occCite")
 ```
+
+## Getting Started
+
+* occCite introduction vignette (https://hannahlowens.github.io/occCite/articles/Advanced.html)
+* occCite advanced feature vignette (https://docs.ropensci.org/rgbif/articles/rgbif.html)
+* Function reference https://hannahlowens.github.io/occCite/reference/index.html
+* YouTube tutorial (https://www.youtube.com/watch?v=7qSCULN_VjY&t=17s)
+
+The `occCite` manuscript is currently in review. Check back soon for more information on that!
 
 ## Meta
 
