+++ README.Rmd
@@ -48,7 +48,7 @@ install.packages("occCite")
 Or, install development version
 
 ```{r eval=FALSE}
-devtools::install_github("DeveloperName/PackageName")("hannahlowens/occCite")
+devtools::install_github("hannahlowens/occCite")
 ```
 
 ```{r}
