### 2021-04-21 796d92a
Remove license reference in description.
