+++ cran-comments.md
@@ -12,7 +12,15 @@ This is a package update. In this version, I have:
 * rhub
 
 ## R CMD check results
-0 errors | 0 warnings | 0 notes | 3 preperrors
+0 errors | 0 warnings | 1 note | 3 preperrors
+
+* A NOTE occurred on x86_64-pc-linux-gnu (64-bit):
+
+Examples with CPU (user + system) or elapsed time > 5s
+                user system elapsed
+studyTaxonList 0.523  0.046   7.083
+
+studyTaxonList() example requires an API response. It is generally complete in < 5s, but this is somewhat dependent on connection and traffic.
 
 * A PREPERROR failure occurred when using rhub to check the Fedora Linux, R-devel, clang, gfortran platform. The message reads: "Error: Bioconductor version '3.13' requires R version '4.1'; R version is too new;". Bioconductor is not a downstream dependency of occCite, so I am unsure how to proceed.
 * A PREPERROR was also signaled for the rhub checks of the Ubuntu Linux 20.04.1 LTS, R-release, GCC platform and Ubuntu Linux 20.04.1 LTS, R-release, GCC platfrom, but the checks still completed successfully with no errors, warnings, or notes.
