+++ NAMESPACE
@@ -19,6 +19,7 @@ exportClasses(occCiteData)
 import(RPostgreSQL)
 import(bib2df)
 import(lubridate)
+importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,"%>%")
 importFrom(dplyr,filter)
 importFrom(ggplot2,aes)
@@ -37,3 +38,4 @@ importFrom(methods,new)
 importFrom(rlang,.data)
 importFrom(stats,complete.cases)
 importFrom(stats,na.omit)
+importFrom(viridis,viridis)
