### 2021-04-27 4268af2
Updated CRAN comments
