### 2021-04-29 e4c2720
Updated development version number, added Zenodo DOI, updated website.
