### 2021-06-11 018650d
Updated to fix error when BIEN did not have citation data for a given dataset key
