+++ NEWS.md
@@ -1,4 +1,4 @@
-# occCite 0.4.7.9000
+# occCite 0.4.8.9000
 
 * Ongoing development
 
