+++ cran-comments.md
@@ -23,10 +23,6 @@ studyTaxonList() example requires an API response. It is generally complete in <
 * A PREPERROR failure occurred when using rhub to check the Fedora Linux, R-devel, clang, gfortran platform. The message reads: "Error: Bioconductor version '3.13' requires R version '4.1'; R version is too new;". Bioconductor is not a downstream dependency of occCite, so I am unsure how to proceed.
 * A PREPERROR was also signaled for the rhub checks of the Ubuntu Linux 20.04.1 LTS, R-release, GCC platform, but the checks still completed successfully with no errors, warnings, or notes.
 
-## CRAN reviewer comments addressed
-
-* LICENSE file deleted
-
 ## Downstream dependencies
 I ran tools::check_packages_in_dir() check on downstream dependencies of 
 occCite. 
