+++ cran-comments.md
@@ -6,22 +6,28 @@ This is a package update. In this version, I have:
 ## Test environments
 * local OS X 10.15.7 install, R 4.0.2
 * win-builder (devel and release)
-* ubuntu 20.04 (devel and release; on GitHub Actions), R 4.0.5
+* ubuntu 20.04 (devel and release; on GitHub Actions), R 4.1.0
+* windows-latest (on GitHub Actions), R 4.1.0
+* mac-latest (on GitHub Actions), R 4.1.0
 * rhub
 
 ## R CMD check results
-0 errors | 0 warnings | 1 note | 2 preperrors
+0 errors | 0 warnings | 2 notes | 1 preperror
 
-* A NOTE occurred on x86_64-pc-linux-gnu (64-bit):
+* A NOTE occurred on win-builder devel:
 
-Examples with CPU (user + system) or elapsed time > 5s
-                user system elapsed
-studyTaxonList 0.523  0.046   7.083
+R Under development (unstable) (2021-06-07 r80458)
 
-studyTaxonList() example requires an API response. It is generally complete in < 5s, but this is somewhat dependent on connection and traffic.
+* A NOTE occurred on windows-latest:
 
-* A PREPERROR failure occurred when using rhub to check the Fedora Linux, R-devel, clang, gfortran platform. The message reads: "Error: Bioconductor version '3.13' requires R version '4.1'; R version is too new;". Bioconductor is not a downstream dependency of occCite, so I am unsure how to proceed.
-* A PREPERROR was also signaled for the rhub checks of the Ubuntu Linux 20.04.1 LTS, R-release, GCC platform, but the checks still completed successfully with no errors, warnings, or notes.
+Maintainer: 'Hannah L. Owens <hannah.owens@gmail.com>'
+  
+Version contains large components (0.4.8)
+
+
+* A PREPERROR failure occurred when using rhub to check the Windows Server 2008 R2 SP1, R-devel, 32/64 bit and Fedora Linux, R-devel, clang, gfortran platforms. The message reads: "Error: Bioconductor does not yet build and check packages for R version 4.2; see https://bioconductor.org/install". Bioconductor is not a downstream dependency of occCite, so I am unsure how to proceed.
+
+* A PREPERROR was signaled when using rhub to check the Ubuntu Linux 20.04.1 LTS, R-release, GCCv platform, but the check still ran successfully.
 
 ## Downstream dependencies
 I ran tools::check_packages_in_dir() check on downstream dependencies of 
