### 2021-06-16 b6d9118
rgbif::name_suggests was returning multiple keys in cases where there were multiple taxonomic authorities
