### 2021-07-22 7c71bb2
Resolving a logic problem that was leading to empty results. Preparing to submit to CRAN.
