### 2021-07-22 0eb31dd
bit64 is needed for tests, not main package.
