### 2021-07-23 4d1ca92
Advanced vignette needs to be leaner.
