+++ R/studyTaxonList.R
@@ -37,7 +37,7 @@
 #' )
 #' phylogeny <- ape::extract.clade(phylogeny, 18)
 #' studyTaxonList(x = phylogeny,
-#'                datasources = c("National Center for Biotechnology Information"))
+#'                datasources = c("GBIF Backbone Taxonomy"))
 #' @export
 studyTaxonList <- function(x = NULL,
                            datasources = "GBIF Backbone Taxonomy") {
