+++ vignettes/Advanced.Rmd
@@ -81,10 +81,10 @@ library(ape)
 #Get tree
 treeFile <- system.file("extdata/Fish_12Tax_time_calibrated.tre", package='occCite')
 phylogeny <- ape::read.nexus(treeFile)
-tree <- ape::extract.clade(phylogeny, 18)
+tree <- ape::extract.clade(phylogeny, 21)
 #Query databases for names
 myPhyOccCiteObject <- studyTaxonList(x = tree, 
-                                     datasources = "National Center for Biotechnology Information")
+                                     datasources = "GBIF Backbone Taxonomy")
 #Query GBIF for occurrence data
 myPhyOccCiteObject <- occQuery(x = myPhyOccCiteObject, 
                             datasources = "gbif",
@@ -104,7 +104,7 @@ plot(myPhyOccCiteObject)
 *or* you can plot the results by species!
 
 ```{r plotting phylogenetic search by species, eval=T, message=FALSE, warning=FALSE, paged.print=FALSE, results='hide', fig.hold='hold', out.width="33%"}
-plot(myPhyOccCiteObject, bySpecies = T)
+plot(myPhyOccCiteObject, bySpecies = T, plotTypes = c("yearHistogram", "source"))
 ```
 
 And then you can print out the citations, separated by species (or not, but in this example, they're separate).
