### 2021-07-23 17eda90
CRAN comments updated
