### 2021-07-26 3a4139f
Updating readme with version and Ecography citation
