### 2021-10-07 ecf7aaf
Cleaning up tests, adding a check for internet connectivity
