+++ tests/testthat/test-occCitePrint.R
@@ -1,0 +1,24 @@
+library(occCite)
+library(ape)
+
+treeFile <- system.file("extdata/Fish_12Tax_time_calibrated.tre", package='occCite')
+phylogeny <- ape::read.nexus(treeFile)
+tree <- ape::extract.clade(phylogeny, 21)
+#Query databases for names
+myPhyOccCiteObject <- studyTaxonList(x = tree,
+                                     datasources = "GBIF Backbone Taxonomy")
+#Query GBIF for occurrence data
+myPhyOccCiteObject <- occQuery(x = myPhyOccCiteObject,
+                               datasources = "gbif",
+                               GBIFDownloadDirectory = system.file('extdata/', package='occCite'),
+                               loadLocalGBIFDownload = T,
+                               checkPreviousGBIFDownload = F)
+myPhyOccCitations <- occCitation(myPhyOccCiteObject)
+
+test_that("regular print", {
+  temp<- print(myPhyOccCitations)
+})
+
+test_that("print by species", {
+  print(myPhyOccCitations, bySpecies = TRUE)
+})
