+++ DESCRIPTION
@@ -51,6 +51,7 @@ Imports:
     htmltools,
     ggplot2,
     rlang,
+    httr,
     tidyr,
     RPostgreSQL,
     RColorBrewer,
@@ -58,5 +59,5 @@ Imports:
     DBI,
     waffle
 VignetteBuilder: knitr
-RoxygenNote: 7.1.1
+RoxygenNote: 7.1.2
 Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
