### 2021-10-07 74319e0
Test for occCite citation print method
