### 2021-10-08 8fd9fc2
Adding checks for internet connectivity, hopefully with more graceful fail.
