+++ R/GBIFLoginManager.R
@@ -60,14 +60,20 @@ GBIFLoginManager <- function(user = NULL, email = NULL, pwd = NULL) {
   pwd <- check_pwd(pwd)
 
   # Test login
-  test <- try(rgbif::occ_download(
-    user = user,
-    email = email,
-    pwd = pwd,
-    rgbif::pred("catalogNumber", 217880)
-  ),
-  silent = T
-  )
+  tryCatch(expr = test <- try(rgbif::occ_download(user = user,
+                                                  email = email,
+                                                  pwd = pwd,
+                                                  rgbif::pred("catalogNumber", 217880)),
+                              silent = T)
+,
+           error = function(e) {
+             message(paste("GBIF unreachable at the moment, please try again later. \n"))
+           })
+
+  if(!exists("test")){
+    return(invisible(NULL))
+  }
+
   if (class(test) != "occ_download") {
     warning("GBIF user login data incorrect.\n")
     return(NULL)
