### 2021-10-08 a569f39
Fixed problem with taxonomic authority gumming up search.
