+++ R/occSum.R
@@ -84,7 +84,7 @@ summary.occCiteData <- function(object, ...) {
 
     for (i in 1:length(x@occResults)) {
       # GBIF counts
-      if (any(x@occSources == "bien",
+      if (any(x@occSources == "bien" && length(x@occSources) == 1,
         is.null(x@occResults[[i]]$GBIF$OccurrenceTable),
         nrow(x@occResults[[i]]$GBIF$OccurrenceTable[!is.na(x@occResults[[i]]$GBIF$OccurrenceTable$DatasetKey), ]) == 0,
         na.rm = T
@@ -101,7 +101,7 @@ summary.occCiteData <- function(object, ...) {
         )
       }
       # BIEN counts
-      if (any(x@occSources == "gbif",
+      if (any(x@occSources == "gbif" && length(x@occSources) == 1,
         is.null(x@occResults[[i]]$BIEN$OccurrenceTable),
         na.rm = T
       )) {
