### 2021-10-12 1a01741
Adding tryCatch to internet search functions
