+++ readme.md
@@ -17,6 +17,8 @@ status](https://badges.ropensci.org/407_status.svg)](https://github.com/ropensci
 
 <!-- badges: end -->
 
+# occCite <img src='man/figures/logo.png' align="right" height="138" />
+
 ## Summary
 
 The `occCite` workflow follows a three-step process. First, the user
@@ -44,8 +46,8 @@ citation(package = "occCite")
     ## Owens H, Merow C, Maitner B, Kass J, Barve V, Guralnick R (2021).
     ## _occCite: Querying and Managing Large Biodiversity Occurrence
     ## Datasets_. doi: 10.5281/zenodo.4726676 (URL:
-    ## https://doi.org/10.5281/zenodo.4726676), R package version 0.4.9, <URL:
-    ## https://CRAN.R-project.org/package=occCite>.
+    ## https://doi.org/10.5281/zenodo.4726676), R package version 0.4.9.9000,
+    ## <URL: https://CRAN.R-project.org/package=occCite>.
     ## 
     ## A BibTeX entry for LaTeX users is
     ## 
@@ -53,7 +55,7 @@ citation(package = "occCite")
     ##     title = {occCite: Querying and Managing Large Biodiversity Occurrence Datasets},
     ##     author = {Hannah Owens and Cory Merow and Brian Maitner and Jamie Kass and Vijay Barve and Robert Guralnick},
     ##     year = {2021},
-    ##     note = {R package version 0.4.9},
+    ##     note = {R package version 0.4.9.9000},
     ##     url = {https://CRAN.R-project.org/package=occCite},
     ##     doi = {10.5281/zenodo.4726676},
     ##   }
