### 2021-10-12 c49e30a
logic error--used all() instead of && in test to fix
