### 2021-10-12 fe6db97
tryCatch for internet-dependant actions
