### 2021-10-12 e1638f5
Error checking taxonomic rectification if there's no internet
