### 2021-10-13 8a0624a
Add a little detail to description RE what exactly gets pulled from aggregators
