### 2021-10-13 9137689
Tests throwing inappropriate error--skip test if cannot connect to globalnames.org
