+++ tests/testthat/test-studyTaxonList.R
@@ -13,6 +13,8 @@ test_that("phylogeny can be read and manipulated as expected", {
 })
 
 test_that("studyTaxonList works with a phylogeny", {
+  skip_if(httr::http_error("https://resolver.globalnames.org/data_sources.json"))
+
   phylogeny <- ape::read.nexus(system.file("extdata/Fish_12Tax_time_calibrated.tre",
     package = "occCite"
   ))
@@ -43,6 +45,8 @@ test_that("studyTaxonList works with a phylogeny", {
 })
 
 test_that("studyTaxonList works with a vector of species", {
+  skip_if(httr::http_error("https://resolver.globalnames.org/data_sources.json"))
+
   taxVector <- c("Buteo buteo", "Buteo buteo hartedi", "Buteo japonicus")
   testResult <- studyTaxonList(x = taxVector, datasources = c("National Center for Biotechnology Information"))
 
