### 2021-10-13 85ae976
more input warning tests
