+++ tests/testthat/test-taxonRectification.R
@@ -43,8 +43,7 @@ test_that("taxonRectification performs as expected", {
   expect_true(testResult$`Best Match`[1] == "Buteo buteo harterti")
   expect_true(testResult$`Searched Taxonomic Databases w/ Matches` == "National Center for Biotechnology Information")
 
-  expect_warning(taxonRectification(
-    taxName = "Buteo buteo hartedi",
-    datasources = "cheese"
-  ))
+  expect_warning(taxonRectification(taxName = "Buteo buteo hartedi", datasources = "cheese"))
+  expect_warning(taxonRectification(taxName = "Buteo buteo hartedi", datasources = NULL))
+  expect_warning(taxonRectification(taxName = "cheese", datasources = "National Center for Biotechnology Information"))
 })
