### 2021-10-29 851e2b1
Slim down file sizes.
