### 2021-10-29 3001aed
Finish 0.5.1 CRAN comments
