### 2021-12-14 #407 @hannahlowens
Hi Laura,
Ok, great! Thanks. I’ve addressed some of their comments, but have been a bit derailed due to some problems with CURL in Linux on CRAN. I will repsond more formally to the review once we’ve chased down the problem (hopefully).

Best,
Hannah
