### 2022-01-13 841f91c
Checks to fail gracefully with informative warnings when servers cannot be reached.
