+++ vignettes/Advanced.R
@@ -29,10 +29,10 @@ library(ape)
 #Get tree
 treeFile <- system.file("extdata/Fish_12Tax_time_calibrated.tre", package='occCite')
 phylogeny <- ape::read.nexus(treeFile)
-tree <- ape::extract.clade(phylogeny, 18)
+tree <- ape::extract.clade(phylogeny, 22)
 #Query databases for names
 myPhyOccCiteObject <- studyTaxonList(x = tree, 
-                                     datasources = "National Center for Biotechnology Information")
+                                     datasources = "GBIF Backbone Taxonomy")
 #Query GBIF for occurrence data
 myPhyOccCiteObject <- occQuery(x = myPhyOccCiteObject, 
                             datasources = "gbif",
@@ -42,11 +42,11 @@ myPhyOccCiteObject <- occQuery(x = myPhyOccCiteObject,
 # What does a multispecies query look like?
 summary(myPhyOccCiteObject)
 
-## ----plotting all species, eval=T, message=FALSE, warning=FALSE, paged.print=FALSE, results='hide', fig.hold='hold', out.width="33%"----
+## ----plotting all species, eval=T, message=FALSE, warning=FALSE, paged.print=FALSE, results='hide', fig.hold='hold', out.width="100%"----
 plot(myPhyOccCiteObject)
 
-## ----plotting phylogenetic search by species, eval=T, message=FALSE, warning=FALSE, paged.print=FALSE, results='hide', fig.hold='hold', out.width="33%"----
-plot(myPhyOccCiteObject, bySpecies = T)
+## ----plotting phylogenetic search by species, eval=T, message=FALSE, warning=FALSE, paged.print=FALSE, results='hide', fig.hold='hold', out.width="100%"----
+plot(myPhyOccCiteObject, bySpecies = T, plotTypes = c("yearHistogram", "source"))
 
 ## ----getting_citations_for_a_multispecies_search, echo=T----------------------
 #Get citations
