+++ NEWS.md
@@ -1,10 +1,7 @@
-# occCite 0.5.1.9000
-
-* Development version
-
 # occCite 0.5.2
 
 * Resubmission after archiving due to dependency archiving.
+* Functions that rely on getting data from servers via an internet connection now behave more gracefully and informatively when the server cannot be reached.
 
 # occCite 0.5.1
 
