+++ pkgdown.yml
@@ -1,8 +1,8 @@
 pandoc: 2.7.3
-pkgdown: 2.0.1
+pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   Advanced: Advanced.html
   Simple: Simple.html
-last_built: 2022-01-03T15:25Z
+last_built: 2022-02-25T20:59Z
 
