+++ vignettes/Advanced.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "Advanced execution of occurrence searches and citations generation"
+title: "Advanced search and citation of occurrences"
 author:
 - Hannah L. Owens
 - Cory Merow
