### 2022-02-28 20a37fb
Added per ROpenSci reviewer comment
