+++ R/getBIENpoints.R
@@ -29,19 +29,19 @@ getBIENpoints <- function(taxon) {
   )
 
   if(!curl::has_internet()){
-    warning("No internet connection available, please try again later. \n")
+    warning("No internet connection available; please try again later. \n")
     return(NULL)
   }
 
   tryCatch(expr = try(occs <- BIEN::BIEN_occurrence_species(species = taxon,
                                                             cultivated = T,
-                                                            only.new.world = F,
+                                                            new.world = F,
                                                             native.status = F,
                                                             collection.info = T,
                                                             natives.only = F),
                       silent = T),
            error = function(e) {
-             message(paste("BIEN unreachable at the moment, please try again later. \n"))
+             message(paste("BIEN unreachable; please try again later. \n"))
            })
 
   if(!exists("occs")){
