+++ R/GBIFLoginManager.R
@@ -63,7 +63,8 @@ GBIFLoginManager <- function(user = NULL, email = NULL, pwd = NULL) {
   tryCatch(expr = test <- try(rgbif::occ_download(user = user,
                                                   email = email,
                                                   pwd = pwd,
-                                                  rgbif::pred("catalogNumber", 217880)),
+                                                  rgbif::pred("catalogNumber",
+                                                              217880)),
                               silent = T),
            error = function(e) {
              message(paste("GBIF unreachable at the moment, please try again later. \n"))
