### 2022-02-28 3810e2a
Code of conduct NOT permissible to CRAN submission
