+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   Advanced: Advanced.html
   Simple: Simple.html
-last_built: 2022-02-28T16:07Z
+last_built: 2022-02-28T19:23Z
 
