### 2022-03-01 f4a814d
Add code of conduct and instructions for contributing.
