+++ README.Rmd
@@ -1,76 +1,0 @@
----
-title: "occCite: Downloading and Citing Biodiversity Data"
-author:
-- Hannah L. Owens
-- Cory Merow
-- Brian Maitner
-- Jamie M. Kass
-- Vijay Barve
-- Robert Guralnick
-date: "`r Sys.Date()`"
-output:
-  md_document:
-    variant: markdown_github
----
-
-```{r setup, include=FALSE}
-library(ape)
-library(occCite)
-knitr::opts_chunk$set(echo = TRUE, error = TRUE)
-knitr::opts_knit$set(root.dir = paste0(path.package("occCite"), "/extdata/"))
-```
-
-<!-- badges: start -->
-[![R build status](https://github.com/hannahlowens/occCite/workflows/R-CMD-check/badge.svg)](https://github.com/hannahlowens/occCite/actions)
-[![cran version](https://www.r-pkg.org/badges/version/occCite)](https://cran.r-project.org/package=occCite)
-[![rstudio mirror downloads](https://cranlogs.r-pkg.org/badges/occCite)](https://github.com/r-hub/cranlogs.app)
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![Codecov test coverage](https://codecov.io/gh/hannahlowens/occCite/branch/main/graph/badge.svg)](https://codecov.io/gh/hannahlowens/occCite?branch=main)
-[![ROpenSci status](https://badges.ropensci.org/407_status.svg)](https://github.com/ropensci/software-review/issues/407)
-[![DOI](https://zenodo.org/badge/151783900.svg)](https://zenodo.org/badge/latestdoi/151783900)
-
-<!-- badges: end -->
-
-# occCite <img src='man/figures/logo.png' align="right" height="138" />
-
-## Summary
-
-The `occCite` workflow follows a three-step process. First, the user inputs one or more taxonomic names (or a phylogeny). `occCite` then rectifies these names by checking them against one or more taxonomic databases, which can be specified by the user (see the [Global Names List](http://gni.globalnames.org/data_sources)). The results of the taxonomic rectification are then kept in an `occCiteData` object in local memory. Next, `occCite` takes the `occCiteData` object and user-defined search parameters to query BIEN (through `rbien`) and/or GBIF(through `rGBIF`) for records. The results are appended to the `occCiteData` object, along with metadata on the search. Finally, the user can pass the `occCiteData` object to `occCitation`, which compiles citations for the primary providers, database aggregators, and `R` packages used to build the dataset.
-
-Please cite occCite. Run the following to get the appropriate citation for the version you're using:
-
-```{r}
-citation(package = "occCite")
-```
-
-## Installation
-
-```{r eval=FALSE}
-install.packages("occCite")
-```
-
-Or, install development version
-
-```{r eval=FALSE}
-devtools::install_github("hannahlowens/occCite")
-```
-
-```{r}
-library("occCite")
-```
-
-## Getting Started
-
-* occCite introduction vignette (https://hannahlowens.github.io/occCite/articles/Simple.html)
-* occCite advanced feature vignette (https://hannahlowens.github.io/occCite/articles/Advanced.html)
-* Function reference https://hannahlowens.github.io/occCite/reference/index.html
-* YouTube tutorial (https://www.youtube.com/watch?v=7qSCULN_VjY&t=17s)
-* Software note in _Ecography_ (https://doi.org/10.1111/ecog.05618)
-
-## Meta
-
-* Please [report any issues or bugs](https://github.com/hannahlowens/occCite/issues).
-* License: GPL-3
-* Get citation information for `occCite` in R using `citation(package = 'occCite')`
-
-- - -
