+++ NAMESPACE
@@ -11,18 +11,22 @@ export(occCiteMap)
 export(occQuery)
 export(prevGBIFdownload)
 export(studyTaxonList)
-export(tabulate.occResults)
 export(taxonRectification)
 exportClasses(GBIFLogin)
 exportClasses(occCiteCitation)
 exportClasses(occCiteData)
 import(RPostgreSQL)
 import(bib2df)
+import(leaflet)
 import(lubridate)
+import(rgbif)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(curl,has_internet)
 importFrom(dplyr,"%>%")
+importFrom(dplyr,bind_rows)
 importFrom(dplyr,filter)
+importFrom(dplyr,mutate)
+importFrom(dplyr,mutate_if)
 importFrom(ggplot2,aes)
 importFrom(ggplot2,element_text)
 importFrom(ggplot2,geom_histogram)
