+++ DESCRIPTION
@@ -14,7 +14,9 @@ Authors@R: c(
   person(given = "Vijay", family = "Barve",,"vijay.barve@gmail.com ",role="aut",
     comment = c(ORCID = "0000-0002-4852-2567")),
   person(given = "Robert P.", family = "Guralnick",,,role="aut",
-    comment = c(ORCID = "0000-0001-6682-1504"))
+    comment = c(ORCID = "0000-0001-6682-1504")),
+  person(given = "Damiano", family = "Oldoni",,,role="rev",
+    comment = c(ORCID = "0000-0003-3445-7562"))
   )
 Author: Hannah L. Owens [aut, cre] (<https://orcid.org/0000-0003-0071-1745>), Cory Merow [aut] (<https://orcid.org/0000-0003-0561-053X>), Brian Maitner [aut] (<https://orcid.org/0000-0002-2118-9880>), Jamie M. Kass [aut] (<https://orcid.org/0000-0002-9432-895X>), Vijay Barve [aut] (<https://orcid.org/0000-0002-4852-2567>), Robert P. Guralnick [aut] (<https://orcid.org/0000-0001-6682-1504>)
 Maintainer: Hannah L. Owens <hannah.owens@gmail.com>
@@ -61,4 +63,3 @@ Imports:
     waffle
 VignetteBuilder: knitr
 RoxygenNote: 7.1.2
-Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
