### 2022-03-01 252e203
Ping GBIF with list request instead of creating a pointless download
