+++ man/occQuery.Rd
@@ -39,7 +39,7 @@ downloaded by the user and stored in the directory specified by
 
 \item{checkPreviousGBIFDownload}{If \code{loadLocalGBIFDownload = T},
 occCite will check for previously-prepared GBIF downloads on the user's
-GBIF account. Setting this option to \code{TRUE} can significantly speed up
+GBIF account. Setting this option to `TRUE` can significantly speed up
 query time if the user has previously queried GBIF for the same taxa.}
 
 \item{options}{A vector of options to pass to \code{\link{occ_download}}.}
