### 2022-03-01 d4604c0
Response to ROpenSci reviewer comments.
