+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 2.7.3
 pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
-  Advanced: Advanced.html
-  Simple: Simple.html
-last_built: 2022-02-28T19:23Z
+  a_Simple: a_Simple.html
+  b_Advanced: b_Advanced.html
+last_built: 2022-03-01T15:57Z
 
