### 2022-03-02 5931019
Reformatted code with usethis::use_tidy_style()
