+++ tests/testthat/test-GBIFLoginManager.R
@@ -5,8 +5,9 @@ library(occCite)
 test_that("GBIFLoginManager error messaging behaves as expected", {
   skip_on_cran()
 
-  test <- try(rgbif::occ_count(country='DK'),
-              silent = T)
+  test <- try(rgbif::occ_count(country = "DK"),
+    silent = T
+  )
   skip_if(class(test) != "numeric", "GBIF connection unsuccessful")
 
   expect_warning(
@@ -20,8 +21,9 @@ test_that("GBIFLoginManager actually picks up login information", {
     "GBIF Login information not available"
   )
 
-  test <- try(rgbif::occ_count(country='DK'),
-              silent = T)
+  test <- try(rgbif::occ_count(country = "DK"),
+    silent = T
+  )
   skip_if(class(test) != "numeric", "GBIF connection unsuccessful")
 
   GBIFLogin <- GBIFLoginManager()
