+++ tests/testthat/test-occCitePrint.R
@@ -1,7 +1,8 @@
 library(occCite)
 library(ape)
 
-treeFile <- system.file("extdata/Fish_12Tax_time_calibrated.tre", package = "occCite")
+treeFile <- system.file("extdata/Fish_12Tax_time_calibrated.tre",
+                        package = "occCite")
 phylogeny <- ape::read.nexus(treeFile)
 tree <- ape::extract.clade(phylogeny, 22)
 # Query databases for names
