+++ tests/testthat/test-studyTaxonList.R
@@ -20,7 +20,8 @@ test_that("studyTaxonList works with a phylogeny", {
     package = "occCite"
   ))
   phylogeny <- ape::extract.clade(phylogeny, 22)
-  testResult <- studyTaxonList(x = phylogeny, datasources = c("National Center for Biotechnology Information"))
+  testResult <- studyTaxonList(x = phylogeny,
+          datasources = c("National Center for Biotechnology Information"))
 
   expect_true(class(testResult) == "occCiteData")
   expect_true("userQueryType" %in% slotNames(testResult))
@@ -31,7 +32,8 @@ test_that("studyTaxonList works with a phylogeny", {
   expect_true("occResults" %in% slotNames(testResult))
 
   expect_true(testResult@userQueryType == "User-supplied phylogeny.")
-  expect_true(testResult@userSpecTaxonomy == "National Center for Biotechnology Information")
+  expect_true(testResult@userSpecTaxonomy ==
+                "National Center for Biotechnology Information")
 
   expect_true(class(testResult@cleanedTaxonomy) == "data.frame")
   expect_true(nrow(testResult@cleanedTaxonomy) == 3)
@@ -50,7 +52,8 @@ test_that("studyTaxonList works with a vector of species", {
   skip_if(httr::http_error("https://resolver.globalnames.org/data_sources.json"))
 
   taxVector <- c("Buteo buteo", "Buteo buteo hartedi", "Buteo japonicus")
-  testResult <- studyTaxonList(x = taxVector, datasources = c("National Center for Biotechnology Information"))
+  testResult <- studyTaxonList(x = taxVector,
+          datasources = c("National Center for Biotechnology Information"))
 
   expect_true(class(testResult) == "occCiteData")
   expect_true("userQueryType" %in% slotNames(testResult))
@@ -61,7 +64,8 @@ test_that("studyTaxonList works with a vector of species", {
   expect_true("occResults" %in% slotNames(testResult))
 
   expect_true(testResult@userQueryType == "User-supplied list of taxa.")
-  expect_true(testResult@userSpecTaxonomy == "National Center for Biotechnology Information")
+  expect_true(testResult@userSpecTaxonomy ==
+            "National Center for Biotechnology Information")
 
   expect_true(class(testResult@cleanedTaxonomy) == "data.frame")
   expect_true(nrow(testResult@cleanedTaxonomy) == 3)
@@ -76,5 +80,6 @@ test_that("studyTaxonList works with a vector of species", {
 })
 
 test_that("studyTaxonList input errors perform as expected", {
-  expect_warning(studyTaxonList(x = 2, datasources = "National Center for Biotechnology Information"))
+  expect_warning(studyTaxonList(x = 2,
+          datasources = "National Center for Biotechnology Information"))
 })
