+++ tests/testthat/test-occSum.R
@@ -15,7 +15,8 @@ test_that("input data object for summary behaves as expected", {
   expect_true(class(myOccCiteObject@cleanedTaxonomy) == "data.frame")
   expect_true("Input Name" %in% colnames(myOccCiteObject@cleanedTaxonomy))
   expect_true("Best Match" %in% colnames(myOccCiteObject@cleanedTaxonomy))
-  expect_true("Taxonomic Databases w/ Matches" %in% colnames(myOccCiteObject@cleanedTaxonomy))
+  expect_true("Taxonomic Databases w/ Matches" %in%
+                colnames(myOccCiteObject@cleanedTaxonomy))
 
   expect_true("occSources" %in% slotNames(myOccCiteObject))
   expect_true(class(myOccCiteObject@occSources) == "character")
@@ -29,7 +30,10 @@ test_that("input data object for summary behaves as expected", {
   expect_true("GBIF" %in% names(myOccCiteObject@occResults[[1]]))
   expect_true(all(c("OccurrenceTable", "Metadata", "RawOccurrences")
   %in% names(myOccCiteObject@occResults[[1]][[1]])))
-  expect_true(all(c("name", "longitude", "latitude", "day", "month", "year", "Dataset", "DatasetKey", "DataService") %in% colnames(myOccCiteObject@occResults[[1]][[1]][[1]])))
+  expect_true(all(c("name", "longitude", "latitude",
+                    "day", "month", "year",
+                    "Dataset", "DatasetKey", "DataService") %in%
+                    colnames(myOccCiteObject@occResults[[1]][[1]][[1]])))
   expect_true("BIEN" %in% names(myOccCiteObject@occResults[[1]]))
   expect_true(all(c("OccurrenceTable", "Metadata", "RawOccurrences")
   %in% names(myOccCiteObject@occResults[[1]][[2]])))
