+++ man/prevGBIFdownload.Rd
@@ -27,8 +27,10 @@ GBIFLogin <- GBIFLoginManager(
   email = "ireneAdler@laScala.org",
   pwd = "sh3r"
 )
-taxKey <- rgbif::name_suggest(q = "Protea cynaroides",
-                              rank = "species")$key[1]
+taxKey <- rgbif::name_suggest(
+  q = "Protea cynaroides",
+  rank = "species"
+)$key[1]
 prevGBIFdownload(
   taxonKey = taxKey,
   GBIFLogin = myGBIFLogin
