### 2022-03-02 d6c9877
Fix tests in cases where there's no login information
