### 2022-03-03 e9fd0e6
Optimizing tests to be more time-efficient on CRAN
