+++ tests/testthat/test-GBIFtableCleanup.R
@@ -2,113 +2,18 @@ context("Testing GBIFtableCleanup")
 
 library(occCite)
 
-test_that("inputs to GBIFtableCleanup from gbifRetriever as expected", {
-  oldwd <- getwd()
-  on.exit(setwd(oldwd))
-
-  test <- try(rgbif::occ_count(country = "DK"),
-    silent = T
-  )
-  skip_if(class(test) != "numeric", "GBIF connection unsuccessful")
-
-  setwd(dir = system.file("extdata/", package = "occCite"))
-  taxon <- "Protea cynaroides"
-  testResult <- occCite:::gbifRetriever(taxon)
-
-  expect_equal(class(testResult[[1]]), "data.frame")
-  expect_equal(names(testResult)[[1]], "OccurrenceTable")
-})
-
-test_that("inputs to GBIFtableCleanup from getGBIFpoints as expected", {
-  skip_on_cran()
-
-  # This is here to make the test robust to internet connectivity and login information
-  GBIFLogin <- try(GBIFLoginManager(), silent = T)
-  skip_if(is(GBIFLogin, "try-error"))
-
-  test <- try(rgbif::occ_download_list(
-    user = GBIFLogin@username,
-    pwd = GBIFLogin@pwd, limit = 1,
-  ),
-  silent = T
-  )
-  skip_if(class(test) != "occ_download", "GBIF login unsuccessful")
-
-  oldwd <- getwd()
-  on.exit(setwd(oldwd))
-
-  setwd(dir = system.file("extdata/", package = "occCite"))
-  taxon <- "Protea cynaroides"
-  if (!dir.exists("temp/")) dir.create("temp/")
-  testResult <- occCite:::getGBIFpoints(
-    taxon = taxon,
-    GBIFLogin = occCite:::GBIFLoginManager(),
-    GBIFDownloadDirectory = "temp/",
-    checkPreviousGBIFDownload = T
-  )
-
-  expect_equal(class(testResult[[1]]), "data.frame")
-  expect_equal(names(testResult)[[1]], "OccurrenceTable")
-
-  unlink("temp/", recursive = T)
-})
-
 test_that("behaves as expected when given a stored GBIF table", {
   oldwd <- getwd()
   on.exit(setwd(oldwd))
 
-  GBIFLogin <- try(GBIFLoginManager(), silent = T)
-  skip_if(is(GBIFLogin, "try-error"))
-
-  test <- try(rgbif::occ_download_list(
-    user = GBIFLogin@username,
-    pwd = GBIFLogin@pwd, limit = 1,
-  ),
-  silent = T
+  test <- try(rgbif::occ_count(country = "DK"),
+              silent = T
   )
-  skip_if(class(test) != "occ_download", "GBIF login unsuccessful")
+  skip_if(class(test) != "numeric", "GBIF connection unsuccessful")
 
   setwd(dir = system.file("extdata/", package = "occCite"))
   taxon <- "Protea cynaroides"
   testResult <- occCite:::gbifRetriever(taxon)[[1]]
-
-  expect_equal(class(testResult), "data.frame")
-  expect_true("gbifID" %in% colnames(testResult))
-  expect_true("name" %in% colnames(testResult))
-  expect_true("longitude" %in% colnames(testResult))
-  expect_true("latitude" %in% colnames(testResult))
-  expect_true("day" %in% colnames(testResult))
-  expect_true("month" %in% colnames(testResult))
-  expect_true("year" %in% colnames(testResult))
-  expect_true("Dataset" %in% colnames(testResult))
-  expect_true("DatasetKey" %in% colnames(testResult))
[…]
