+++ tests/testthat/test-getBIENpoints.R
@@ -5,10 +5,6 @@ library(occCite)
 test_that("BIEN behaves as expected", {
   skip_on_cran()
   skip_if(!curl::has_internet(), "internet connection unsuccessful")
-  test <- try(rgbif::occ_count(country = "DK"),
-    silent = T
-  )
-  skip_if(class(test) != "numeric", "GBIF connection unsuccessful")
 
   testResult <- BIEN::BIEN_occurrence_species(
     species = "Protea cynaroides",
@@ -31,6 +27,9 @@ test_that("BIEN behaves as expected", {
 test_that("getBIENpoints behaves as expected", {
   skip_on_cran()
   skip_if(!curl::has_internet(), "internet connection unsuccessful")
+
+  expect_error(getBIENpoints())
+
   testResult <- getBIENpoints(taxon = "Protea cynaroides")
   expect_equal(class(testResult), "list")
   expect_equal(length(testResult), 3)

[…]
