### 2022-03-03 a39d6ae
Switch example to dontrun
