+++ NEWS.md
@@ -1,3 +1,7 @@
+# occCite 0.5.2.9000
+
+* In development
+
 # occCite 0.5.2
 
 * Resubmission after archiving due to dependency archiving.
