### 2022-03-04 #407 @hannahlowens
# Review Response

Hi @ldecicco-USGS and @damianooldoni,

After a long time (sincere apologies there), I have worked through Damiano's wonderfully thorough and helpful review, and I have added him as a reviewer in the package citation. Below, I will go through his comments, point-by-point.

## Regarding "points"

At several places in the review, the reviewer suggests replacing "point(s)" with "occurrence(s)", since it's a bit more intuitive and specific. I have done this throughout the documentation, although argument and function names remain the same. There are already downstream dependencies that use the existing names. I will work with the developers of those packages to transition from "points" to "occurrences" as a natural part of our ongoing development collaboration.

## README

-   README.md has been reworked to include sections on installing and citing the package
-   The demonstration of usage has been replaced with links to "Simple" and "Advanced" vignettes
-   Requested badges have been added, as has a Zenodo DOI badge
-   Code of conduct and contribution guidelines have been added and are linked in the README
-   The Global Names List link has been fixed
-   Good catch on the `occ_search()` hard limit. I have fixed information throughout where appropriate.

## Vignettes

-   Vignettes are now split into "Simple" and "Advanced" workflows.
-   Titles have also been shortened: "Simple search and citation of occurrences" and "Advanced search and citation of occurrences".
-   Code chunks all work without error now.
-   The "no date" warning has been fixed. As a citation package, that one was for sure embarrassing.

## Code Style

-   Reformatted using use_tidy_style() and broke up lines longer than 80 characters when it did not strongly and adversely affect readability or functionality.

## Functions

### `GBIFLoginManager()`

-   Silent download trigger removed--excellent point. Instead, function pings GBIF API for a list of previous downloads. This is done to 1) check connectivity and fail gracefully if the server is unresponsive and 2) check the user's login credentials are correct.
-   Typo fixed

### `getGBIFpoints()`

-   GBIF now mentioned in the description
-   Now only retrieves presences (Lines 98-100). This was another excellent catch. I had not thought absences would be included, but of course GBIF doesn't just serve museum data (my background), they also serve all kinds of other biodiversity data. Thank you! I've added handling custom predicates as an issue to be addressed in future development. There are some complexities there I need to think carefully about. I've added this as an issue on the gitHub. At any rate, the automatic predicates are now stipulated in @details.
-   Handling for higher taxonomies and hemihomonies has also been added as an issue on gitHub. The potential for multiple taxonomic keys of which the first is not correct is potentially a problem.
-   This function was written primarily internally for use by `occQuery()`, which passes the function an already validated species name (or a name that has been extracted from a phylogeny and then validated), hence the documentation for @value x being what it is.
-   The package was designed for use particularly in species distribution modeling/ecological niche modeling workflows, hence the taxonomic focus. I recognize that it could potentially be useful in a more geographic context, but we first want to really nail the species-level taxonomic-focused functionality to establish the package first. -Typo at Line 34 fixed.

### `prevGBIFdownload()`

-   There is no reason to specify a default value of "No key" for `taxonKey`. This was an inelegant early solution that was never removed. It's now `NULL`.
-   I really like the suggestion of adding the ability to pass `prevGBIFdownload()` a specific download key so that the user can bring searches started elsewhere into the `occCite` workflow. I have added this as an issue, as I think it will take some careful thought to integrate fully and intuitively.

### `occCitation()`

-   `RPostgreSQL` is imported whole since the list of functions being imported was running rather long.

### `tabulate.occResults()`

-   Fixed so `@keywords internal` and added `@noRd`
-   Functions listed now imported from `dplyr` via `@importFrom`
-   `filter` removed from `@importFrom dplyr`

### `occCiteMap()`

-   `leaftlet` is imported whole since the list of functions being imported was running rather long.

### All other functions

Documentation check done. Edited for clarity and completeness.

## Tests

-   The `catalogNumber` search was a clunky solution to testing GBIF API connectivity, which I have now replaced with an `occ_download_list()`query. These tests are necessary to confirm connectivity with the GBIF API when running the tests, and to flag errors that might arise if there are changes in how `rgbif` is structured. Currently, as `occCite`'s database scope is limited to GBIF and BIEN, the lack of functionality in either package bringing down `occCite` is desirable. In the future when the scope of the package is expanded, making is less dependent on `rgbif` will be a higher priority.
-   I have fixed the skipped tests--`GBIFLogin()` should have been run to collect login information.
-   The `try()` elements are included to make the package tests robust to situations where there is no ability to connect to the APIs. This is a CRAN requirement.

# Conclusion

Thank you again. I firmly believe this review was incredibly helpful and made `occCite` more robust, useful, and clear. Please let me know how to proceed. Is a second review still coming?
