### 2022-03-15 #407 @maelle
:wave: @hannahlowens  Is any help needed with the repo transfer?
I'm asking as the package will be listed as new package in Friday's newsletter if it's transferred by then. :smile_cat:
