### 2022-03-17 #407 @maelle
@hannahlowens can you please try again with `@ropensci-review-bot finalize transfer of occCite`? Thank you!
